% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.Wald_lmm}
\alias{aggregate.Wald_lmm}
\title{Pool Estimates}
\usage{
\method{aggregate}{Wald_lmm}(
  x,
  method,
  rhs = NULL,
  columns = NULL,
  qt = NULL,
  level = 0.95,
  df = NULL,
  tol = 1e-10,
  ...
)
}
\arguments{
\item{x}{output of \code{anova.lmm}, \code{rbind.Wald_lmm}, or \code{mlmm}.}

\item{method}{[character] method(s) used to pool the estimates}

\item{rhs}{[logical or numeric vector] either the right-hand side of the null hypotheses to be tested
or whether the value of the summary statistic under the null hypothesis should be computed - mostly relevant when \code{method="p.rejection"} where this can be time consuming.}

\item{columns}{[character] column(s) to be exported}

\item{qt}{[numeric or character] critical quantile to be considered when evaluating the proportion of rejected hypotheses.
Can also be the name of a multiple testing method from which the quantile should be computed.}

\item{level}{[numeric, 0-1] nominal coverage of the confidence intervals.}

\item{df}{[logical] Should a Student's t-distribution be used to model the distribution of the summary statistic. Otherwise a normal distribution is used.}

\item{tol}{[numeric, >0] threshold below which a pseudo-inverse is used when inverted a matrix, i.e., the eigen-values are truncated.}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Pool estimates from one or several linear mixed models. \code{confint} is the prefered method as this is meant for internal use.
}
\details{
Use a first order delta method to evaluate the standard error.
When \code{method="p.rejection"} the p-value and distribution under the null is evaluated by simulating data using a Gaussian or t-copula, which can be time consumming.
The number of sample is controlled by the argument \code{n.sampleCopula} in \code{\link{LMMstar.options}}.
}
\seealso{
\code{\link{confint.Wald_lmm}}, \code{\link{confint.rbindWald_lmm}}, \code{\link{confint.mlmm}}
}
