% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allCoef.R
\name{allCoef}
\alias{allCoef}
\title{Extract Model Coefficients With Confidence Intervals}
\usage{
allCoef(object, conf.int, conf.level, effects)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme}, or \code{lmm} object.}

\item{conf.int}{[logical] Should the confidence interval be output}

\item{conf.level}{[numeric 0-1] Confidence level of the confidence intervals.}

\item{effects}{[character vector] Type of coefficient to be output.
Can be coefficients relative to the expectation of the outcome (\code{"mean"})
or to the variance-covariance structure of the residuals (\code{"variance"}).}
}
\description{
Extract all model coefficients with confidence intervals.
}
\examples{
data(gastricbypassL)
library(nlme)

#### linear model ####
## (wrong model as it does not account for repeated measurements)
e.lm <- lm(weight ~ time, data = gastricbypassL)
allCoef(e.lm)

#### gls model ####
e.gls <- gls(weight ~ time,
             correlation = corSymm(form =~as.numeric(visit)|id),
             weights = varIdent(form =~1|visit),
             data = gastricbypassL)
allCoef(e.gls)

#### lme model ####
e.lme <- lme(weight ~ time,
             random = ~1|id,
             weights = varIdent(form =~1|visit),
             data = gastricbypassL)
allCoef(e.lme)

#### lmm model ####
e.lmm <- lmm(weight ~ time, covariance = ~visit|id, data = gastricbypassL)
allCoef(e.lmm)
}
