% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.lmm}
\alias{anova.lmm}
\title{Multivariate Tests For Linear Mixed Model}
\usage{
\method{anova}{lmm}(
  object,
  effects = NULL,
  rhs = NULL,
  robust = NULL,
  df = NULL,
  univariate = TRUE,
  multivariate = TRUE,
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  simplify = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{lmm} object. Only relevant for the anova function.}

\item{effects}{[character or numeric matrix] Should the Wald test be computed for all variables (\code{"all"}),
or only variables relative to the mean (\code{"mean"} or \code{"fixed"}),
or only variables relative to the variance structure (\code{"variance"}),
or only variables relative to the correlation structure (\code{"correlation"}).
Can also be use to specify linear combinations of coefficients or a contrast matrix, similarly to the \code{linfct} argument of the \code{multcomp::glht} function.}

\item{rhs}{[numeric vector] the right hand side of the hypothesis. Only used when the argument \code{effects} is a matrix.}

\item{robust}{[logical] Should robust standard errors (aka sandwich estimator) be output instead of the model-based standard errors. 
Can also be \code{2} compute the degrees of freedom w.r.t. robust standard errors instead of w.r.t. model-based standard errors.}

\item{df}{[logical] Should degrees of freedom be estimated using a Satterthwaite approximation?
If yes F-distribution (multivariate) and Student's t-distribution (univariate) are used.
Other chi-squared distribution and normal distribution are used.}

\item{univariate}{[logical] Should an estimate, standard error, confidence interval, and p-value be output for each hypothesis?}

\item{multivariate}{[logical] Should all hypotheses be simultaneously tested using a multivariate Wald test?}

\item{transform.sigma, transform.k, transform.rho}{are passed to the \code{vcov} method. See details section in \code{\link{coef.lmm}}.}

\item{simplify}{[logical] should only the estimates, variance-covariance with its gradient, and degrees of freedom relative to the parameters involved in the Wald test be stored (TRUE)
or relative to all model parameters (0.5) along with their iid decomposition (FALSE).}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A data.frame (LRT) or a list of containing the following elements (Wald):\itemize{
\item \code{multivariate}: data.frame containing the multivariate Wald test.
The column \code{df.num} refers to the degrees of freedom for the numerator (i.e. number of hypotheses)
wherease the column \code{df.denum} refers to the degrees of freedom for the denominator (i.e. Satterthwaite approximation).
\item \code{univariate}: data.frame containing each univariate Wald test.
\item \code{glht}: used internally to call functions from the multcomp package.
\item \code{object}: list containing key information about the linear mixed model.
\item \code{args}: list containing argument values from the function call.
}
}
\description{
Simultaneous tests of linear combinations of the model paramaters using Wald tests or Likelihood Ratio Test (LRT).
}
\details{
By default adjustment of confidence intervals and p-values for multiple comparisons is based on the distribution of the maximum-statistic.
This is refered to as a single-step Dunnett multiple testing procedures in table II of Dmitrienko et al. (2013).
It is performed using the multcomp package with the option \code{test = adjusted("single-step")} with equal degrees of freedom
or by simulation using a Student's t copula with unequal degrees of freedom (more in the note of the details section of \code{\link{confint.Wald_lmm}}).
}
\examples{
#### simulate data in the long format ####
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")

#### fit Linear Mixed Model ####
eUN.lmm <- lmm(Y ~ visit + X1 + X2 + X5,
               repetition = ~visit|id, structure = "UN", data = dL)

#### Multivariate Wald test ####
## F-tests
anova(eUN.lmm)
anova(eUN.lmm, effects = "all")
anova(eUN.lmm, robust = TRUE, df = FALSE)
summary(anova(eUN.lmm), method = "bonferroni")

## user defined F-test
summary(anova(eUN.lmm, effects = c("X1=0","X2+X5=10")))

## chi2-tests
anova(eUN.lmm, df = FALSE)

## with standard contrast
if(require(multcomp)){
amod <- lmm(breaks ~ tension, data = warpbreaks)
e.amod <- anova(amod, effect = mcp(tension = "Tukey"))
summary(e.amod)
}

#### Likelihood ratio test ####
eUN0.lmm <- lmm(Y ~ X1 + X2, repetition = ~visit|id, structure = "UN", data = dL)
anova(eUN.lmm, eUN0.lmm) 

eCS.lmm <- lmm(Y ~ X1 + X2 + X5, repetition = ~visit|id, structure = "CS", data = dL)
anova(eUN.lmm, eCS.lmm)
}
\references{
Dmitrienko, A. and D'Agostino, R., Sr (2013), Traditional multiplicity adjustment methods in clinical trials. Statist. Med., 32: 5172-5218. https://doi.org/10.1002/sim.5990.
}
\seealso{
\code{\link{summary.Wald_lmm}} or \code{\link{confint.Wald_lmm}} for a summary of the results. \cr
\code{\link{autoplot.Wald_lmm}} for a graphical display of the results. \cr
\code{\link{rbind.Wald_lmm}} for combining result across models and adjust for multiple comparisons. \cr
}
\keyword{htest}
