% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\alias{anova.lmm}
\title{Multivariate Wald Tests For Linear Mixed Models}
\usage{
\method{anova}{lmm}(
  object,
  effects = "all",
  df = !is.null(object$df),
  print = TRUE,
  print.null = FALSE,
  type.object = "lmm",
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  transform.names = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{effects}{[character] Should the Wald test be computed for all variables (\code{"all"}),
or only variables relative to the mean (\code{"mean"}),
or only variables relative to the variance structure (\code{"variance"}),
or only variables relative to the correlation structure (\code{"correlation"}).
Can also be use to specify linear combinations of coefficients, similarly to the \code{linfct} argument of the \code{multcomp::glht} function.}

\item{df}{[logical] Should a F-distribution be used to model the distribution of the Wald statistic. Otherwise a chi-squared distribution is used.}

\item{print}{[logical] should the output be printed in the console.}

\item{print.null}{[logical] should the null hypotheses be printed in the console.}

\item{type.object}{[character] Set this argument to \code{"gls"} to obtain the output from the gls object and related methods.}

\item{...}{Not used. For compatibility with the generic method.}

\item{type.information, transform.sigma, transform.k, transform.rho, transform.names}{are passed to the \code{vcov} method. See details section in \code{\link{coef.lmm}}.}
}
\value{
A list of matrices containing the following columns:\itemize{
\item \code{null}: null hypothesis
\item \code{statistic}: value of the test statistic
\item \code{df.num}: degrees of freedom for the numerator (i.e. number of hypotheses)
\item \code{df.denom}: degrees of freedom for the denominator (i.e. Satterthwaite approximation)
\item \code{p.value}: p-value.
}
as well as an attribute contrast containing the contrast matrix encoding the linear combinations of coefficients (in columns) for each hypothesis (in rows).
}
\description{
Perform a Wald test testing simultaneously several null hypotheses corresponding to linear combinations of the model paramaters.
}
\examples{
## simulate data in the long format
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")

## fit mixed model
eUN.lmm <- lmm(Y ~ X1 + X2 + X5, repetition = ~visit|id, structure = "UN", data = dL, df = FALSE)

## chi-2 test
anova(eUN.lmm)

anova(eUN.lmm, effects = c("X1=0","X2+X5=0"))

## F-test
\dontrun{
anova(eUN.lmm, df = TRUE)
}

}
