% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approxAUC}
\alias{approxAUC}
\title{AUC: numerical integral of a function}
\usage{
approxAUC(
  x,
  y,
  from,
  to,
  method = "trapezoid",
  subdivisions = 100,
  name = "AUC",
  na.rm = FALSE
)
}
\arguments{
\item{x}{[numeric vector] x-values.}

\item{y}{[numeric vector] image of the x-values by the function.}

\item{from}{[numeric] lower border of the intergration domain}

\item{to}{[numeric] upper border of the intergration domain}

\item{method}{[character] the type of interpolation: \code{"trapezoid"}, \code{"step"} or \code{"spline"}.}

\item{subdivisions}{[integer] number of subdivisions to be used when performing numerical integration of the spline.
Only relevant when \code{method="spline"}.}

\item{name}{[character] how to name the output. Can be set to \code{NULL} or \code{FALSE} to output a numeric without name.}

\item{na.rm}{[logical] in presence of missing values, should complete.cases of \code{x} and \code{y} will be used?}
}
\value{
a numeric value.
}
\description{
Evaluate the Area Under the Curve (AUC) based on discrete observations y = f(x),
using interpolation of order 0 (step), 1 (trapezoidal), or 3 (natural cubic splines).
The AUC is the integral of a function over an interval [from,to].
}
\details{
This function is a simplified version of the \code{AUC} function from the DescTools package.
}
\examples{
## same examples as DescTools::AUC
approxAUC(x=c(1,3), y=c(1,1), from = 1, to = 3)

approxAUC(x=1:3, y=c(1,2,4), from = 1, to = 3)
approxAUC(x=1:3, y=c(1,2,4), from = 1, to = 3, method = "step")

x <- seq(0, pi, length.out=200)
approxAUC(x=x, y=sin(x), from = 0, to = pi)
approxAUC(x=x, y=sin(x), from = 0, to = pi, method = "spline")

}
