% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/plot.R
\name{autoplot.correlate}
\alias{autoplot.correlate}
\alias{plot.correlate}
\title{Graphical Display For Correlation Matrix}
\usage{
\method{autoplot}{correlate}(
  object,
  index,
  size.text = 16,
  name.legend = "Correlation",
  title = NULL,
  scale = ggplot2::scale_fill_gradient2,
  type.cor = "both",
  low = "blue",
  high = "red",
  mid = "white",
  midpoint = mean(limits),
  limits = c(-1, 1),
  ...
)

\method{plot}{correlate}(x, ...)
}
\arguments{
\item{object}{[correlate] list of list of correlation matrix}

\item{index}{[character vector] optional vector used to select some of the correlation matrix.}

\item{size.text}{[numeric, >0] size of the font used to display text.}

\item{name.legend}{[character] title for the color scale.}

\item{title}{[character] title for the graph.}

\item{scale}{[function] color scale used for the correlation.}

\item{type.cor}{[character] should the whole correlation matrix be displayed (\code{"both"}),
or only the element in the lower or upper triangle (\code{"lower"}, \code{"upper"}).}

\item{low}{[character] color used to represent low correlation.}

\item{high}{[character] color used to represent high correlation.}

\item{mid}{[character] color used to represent moderate correlation.}

\item{midpoint}{[numeric] value defining a modereate correlation.}

\item{limits}{[numeric vector of length 2] values defining a low and high correlation.}
}
\description{
Graphical representation for correlation matrix
}
\section{Functions}{
\itemize{
\item \code{plot(correlate)}: Graphical Display For Correlation Matrix

}}
