% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/plot.R
\name{autoplot.summarizeNA}
\alias{autoplot.summarizeNA}
\alias{plot.summarizeNA}
\title{Graphical Display of Missing Data Pattern}
\usage{
\method{autoplot}{summarizeNA}(
  object,
  variable = NULL,
  size.text = 16,
  add.missing = " missing",
  order.pattern = NULL,
  decreasing.order = FALSE,
  title = NULL,
  labeller = "label_value",
  ...
)

\method{plot}{summarizeNA}(x, ...)
}
\arguments{
\item{object, x}{a \code{summarizeNA} object, output of the \code{\link{summarizeNA}} function.}

\item{variable}{[character] variable for which the missing patterns should be displayed.
Only required when the argument \code{repetition} has been specified when calling \code{summarizeNA}.}

\item{size.text}{[numeric, >0] size of the font used to display text.}

\item{add.missing}{[logical] should the number of missing values per variable be added to the x-axis tick labels.}

\item{order.pattern}{[numeric vector or character] in which order the missing data pattern should be displayed. Can either be a numeric vector indexing the patterns or a character refering to order the patterns per number of missing values (\code{"n.missing"}) or number of observations (\code{"frequency"}).}

\item{decreasing.order}{[logical] when ordering the missing data pattern (see argument \code{order.pattern}) should the sort order be decreasing?
Passed to \code{base::order}.}

\item{title}{[character] title of the graphical display. Passed to \code{ggplot2::ggtitle}.}

\item{labeller}{[character] how to label each facet: only the value (\code{"label_value"}) or with also the variable name (\code{"label_both"}). Passed to \code{ggplot2::facet_wrap}.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A list with two elements \itemize{
\item \code{data}: data used to create the graphical display.
\item \code{plot}: ggplot object.
}
}
\description{
Graphical representation of the possible missing data patterns in the dataset.
}
\section{Functions}{
\itemize{
\item \code{plot(summarizeNA)}: Graphical Display of Missing Data Pattern

}}
\keyword{hplot}
