% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufit.R
\name{autoplot.ufit}
\alias{autoplot.ufit}
\title{Display Unique Fitted Values}
\usage{
\method{autoplot}{ufit}(
  object,
  x = NULL,
  sep.strata = ", ",
  geom_confint = "errorbar",
  size.point = 3,
  size.line = 1.75,
  size.ci = 1,
  alpha.ribbon = 0.5
)
}
\arguments{
\item{object}{output of the \code{ufit} function.}

\item{x}{[character vector] time variable.}

\item{sep.strata}{[character] symbol used to separate the strata values in the caption}

\item{geom_confint}{[character] geometry used to display the confidence intervals.
Can be \code{"none"} for no display, \code{"errorbar"} to use intervals, or \code{"ribbon"} to use a shaded area.}

\item{size.point}{[numeric >0] size argument used in \code{geom_point} to display point estimates.}

\item{size.line}{[numeric >0] size argument used in \code{geom_line} to connect point estimates belonging to the same strata.}

\item{size.ci}{[numeric >0] size argument used in \code{geom_errorbar} to display the confidence intervals.}

\item{alpha.ribbon}{[numeric 0-1] transparency parameter used to display the confidence intervals when using \code{geom_confint = "ribbon"}.}
}
\description{
Display all the possible fitted values among the possible combinaisons of categorical covariables.
Continuous covariates are set to a single value.
}
