% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.lmm}
\alias{autoplot.lmm}
\alias{autoplot.Wald_lmm}
\alias{autoplot.partialCor}
\title{Graphical Display For Linear Mixed Models}
\usage{
\method{autoplot}{lmm}(
  object,
  obs.alpha = 0,
  obs.size = c(2, 0.5),
  at = NULL,
  time.var = NULL,
  color = TRUE,
  ci = TRUE,
  ci.alpha = NA,
  plot = TRUE,
  mean.size = c(3, 1),
  size.text = 16,
  position.errorbar = "identity",
  ...
)

\method{autoplot}{Wald_lmm}(
  object,
  type = "forest",
  plot = TRUE,
  size.text = 16,
  add.args = NULL,
  ...
)

\method{autoplot}{partialCor}(
  object,
  plot = TRUE,
  size.text = 16,
  limits = c(-1, 1.00001),
  low = "blue",
  mid = "white",
  high = "red",
  midpoint = 0,
  ...
)
}
\arguments{
\item{object}{a \code{partialCor} object.}

\item{obs.alpha}{[numeric, 0-1] When not NA, transparency parameter used to display the original data by cluster.}

\item{obs.size}{[numeric vector of length 2] size of the point and line for the original data.}

\item{at}{[data.frame] values for the covariates at which to evaluate the fitted values.}

\item{time.var}{[character] x-axis variable for the plot.}

\item{color}{[character] name of the variable in the dataset used to color the curve.}

\item{ci}{[logical] should confidence intervals be displayed?}

\item{ci.alpha}{[numeric, 0-1] When not NA, transparency parameter used to display the confidence intervals.}

\item{plot}{[logical] should the plot be displayed?}

\item{mean.size}{[numeric vector of length 2] size of the point and line for the mean trajectory.}

\item{size.text}{[numeric, >0] size of the font used to display text.}

\item{position.errorbar}{[character] relative position of the errorbars.}

\item{...}{Not used. For compatibility with the generic method.}

\item{type}{[character] what to display: a forest plot (\code{"forest"}) or a heatmap (\code{"heat"}).}

\item{add.args}{[list] additional arguments used to customized the graphical display.
Must be a named list. See details.}

\item{limits}{[numeric vector of length 2] minimum and maximum value of the colorscale relative to the correlation.}

\item{low, mid, high:}{[character] color for the the colorscale relative to the correlation.}

\item{midpoint:}{[numeric] correlation value associated with the color defined by argument \code{mid}.}
}
\value{
A list with two elements \itemize{
\item \code{data}: data used to create the graphical display.
\item \code{plot}: ggplot object.
}

A list with two elements \itemize{
\item \code{data}: data used to create the graphical display.
\item \code{plot}: ggplot object.
}

A list with two elements \itemize{
\item \code{data}: data used to create the graphical display.
\item \code{plot}: ggplot object.
}
}
\description{
Extract and display the correlation modeled via the linear mixed model.
}
\details{
Argument \strong{add.args}: parameters specific to the forest plot: \itemize{
\item \code{color}: [logical] should the estimates be colored by global null hypothesis, e.g. when testing the effect of a 3 factor covariate, the two corresponding coefficient will have the same color.
\item \code{ci}: [logical] should confidence intervals be displayed?
\item \code{size.estimate}: [numeric, >0] size of the dot used to display the estimates.
\item \code{size.ci}: [numeric, >0] thickness of the line used to display the confidence intervals.
\item \code{width.ci}: [numeric, >0] width of the line used to display the confidence intervals.
\item \code{size.null}: [numeric, >0] thickness of the line used to display the null hypothesis. 
}
Parameters specific to the heatmap plot: \itemize{
\item \code{limits}: [numeric vector of length 2] minimum and maximum value of the colorscale relative to the correlation.
\item \code{low}, \code{mid}, \code{high}: [character] color for the the colorscale relative to the correlation.
\item \code{midpoint}: [numeric] correlation value associated with the color defined by argument \code{mid}
}
}
\examples{
## From the multcomp package
if(require(datasets) && require(ggplot2)){

## only tests with 1 df
e.lmm <- lmm(Fertility ~ Agriculture + Examination + Education + Catholic + Infant.Mortality, data = swiss)
e.aovlmm <- anova(e.lmm)

autoplot(e.aovlmm, type = "forest")
autoplot(e.aovlmm, type = "heat") ## 3 color gradient
autoplot(e.aovlmm, type = "heat", add.args = list(mid = NULL)) ## 2 color gradient

## test with more than 1 df
e.lmm2 <- lmm(breaks ~ tension + wool, data = warpbreaks)
e.aovlmm2 <- anova(e.lmm2)
autoplot(e.aovlmm2)
autoplot(e.aovlmm2, add.args = list(color = FALSE))
}

}
