% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ckdL}
\alias{ckdL}
\title{CKD long}
\usage{
data(ckdL)
}
\description{
TODO

\itemize{
\item id Patient identifier
\item allocation
\item sex
\item age
\item visit
\item time
\item pwv
\item aix
\item dropout
}
}
\references{
TO ADD
}
\keyword{data}
