% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\alias{collapse.data.frame}
\alias{collapse.list}
\alias{collapse.matrix}
\title{Aggregate Dataset by Columns}
\usage{
collapse.data.frame(object, sep = ".", as.factor = TRUE, ...)

collapse.list(object, sep = ".", as.factor = TRUE, ...)

collapse.matrix(object, ...)
}
\arguments{
\item{object}{dataset}

\item{sep}{[character] character strings used to concatenate values from different columns.}

\item{as.factor}{[logical] should the result be converted into a factor or be numeric or character.}

\item{...}{Not used. For compatibility with the generic method.½}
}
\description{
Alternative to \code{\link{interaction}} (which can be quite slow) to aggregate columns of a dataset.
}
\examples{
df1 <- data.frame(var1 = 1:10)
nlme::collapse(df1)

df2 <- data.frame(var1 = 1:10, var2 = -(1:10))
nlme::collapse(df2)
interaction(df2, sep = ".", drop = TRUE)

df3 <- data.frame(var1 = 1:1000, var2 = -(1:1000), var3 = (1:1000)/1000)
test <- nlme::collapse(df3)
GS <- interaction(df3, sep = ".", drop = TRUE)
}
\keyword{internal}
