% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.anova_lmm.R
\name{confint.anova_lmm}
\alias{confint.anova_lmm}
\title{Confidence Intervals for Multivariate Wald Tests}
\usage{
\method{confint}{anova_lmm}(
  object,
  parm,
  level = 0.95,
  method = "single-step",
  simplify = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{anova_lmm} object}

\item{parm}{Not used. For compatibility with the generic method.}

\item{level}{[numeric, 0-1] nominal coverage of the confidence intervals.}

\item{method}{[character] type of adjustment for multiple comparisons: one of \code{"none"}, \code{"bonferroni"}, \code{"single-step"}.}

\item{simplify}{[logical] Return a data.frame instead of a list containing a data.frame when possible.}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Compute confidence intervals for linear hypothesis tests, possibly with adjustment for multiple comparisons.
}
