% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.rbindWald_lmm}
\alias{confint.rbindWald_lmm}
\title{Confidence Intervals From Combined Wald Tests Applied to Linear Mixed Models}
\usage{
\method{confint}{rbindWald_lmm}(
  object,
  parm,
  level = 0.95,
  df = NULL,
  method = NULL,
  columns = NULL,
  ordering = NULL,
  backtransform = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{rbindWald_lmm} object.}

\item{parm}{Not used. For compatibility with the generic method.}

\item{level}{[numeric, 0-1] nominal coverage of the confidence intervals.}

\item{df}{[logical] Should a Student's t-distribution be used to model the distribution of the Wald statistic. Otherwise a normal distribution is used.}

\item{method}{[character] Should pointwise confidence intervals be output (\code{"none"}) or simultaneous confidence intervals (\code{"bonferroni"}, ..., \code{"fdr"}, \code{"single-step"}, \code{"single-step2"}) and/or confidence intervals for pooled linear contrast estimates (\code{"average"}, \code{"pool.se"}, \code{"pool.gls"}, \code{"pool.gls1"}, \code{"pool.rubin"}, \code{"p.rejection"})?
Only relevant when \code{effects = "Wald"}.}

\item{columns}{[character vector] Columns to be output.
Can be any of \code{"estimate"}, \code{"se"}, \code{"statistic"}, \code{"df"}, \code{"null"}, \code{"lower"}, \code{"upper"}, \code{"p.value"}.}

\item{ordering}{[character] should the output be ordered by name of the linear contrast (\code{"contrast"}) or by model (\code{"model"}).}

\item{backtransform}{[logical] should the estimates be back-transformed?}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Combine pointwise or simultaneous confidence intervals relative to linear contrasts of parameters from different linear mixed models. 
 Can also output p-values (corresponding to pointwise confidence intervals) or adjusted p-values (corresponding to simultaneous confidence intervals).
}
\details{
Argument \bold{method}: the following pooling are available:
\itemize{
 \item \code{"average"}: average estimates
 \item \code{"pool.se"}: weighted average of the estimates, with weights being the inverse of the squared standard error. 
 \item \code{"pool.gls"}: weighted average of the estimates, with weights being based on the variance-covariance matrix of the estimates. When this matrix is singular, the Moore–Penrose inverse is used which correspond to truncate the spectral decomposition for eigenvalues below \eqn{10^{-12}}. 
 \item \code{"pool.gls1"}: similar to \code{"pool.gls"} with weights shrinked toward the average whenever they exceed 1 in absolute value.
 \item \code{"pool.rubin"}: average of the estimates and compute the uncertainty according to Rubin's rule (Barnard et al. 1999). Validity requires the congeniality condition of Meng (1994).
 \item \code{"p.rejection"}: proportion of null hypotheses where there is evidence for an effect. By default the critical quantile (defining the level of evidence required) is evaluated using a \code{"single-step"} method but this can be changed by adding adjustment method in the argument \code{method}, e.g. \code{effects=c("bonferronin","p.rejection")}.
}
}
