% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint}
\alias{confint}
\alias{confint.lmm}
\title{Statistical Inference for Linear Mixed Model.}
\usage{
\method{confint}{lmm}(
  object,
  parm = NULL,
  effects = "all",
  level = 0.95,
  null = NULL,
  type.object = "lmm",
  strata = NULL,
  df = !is.null(object$df),
  type.information = NULL,
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  transform.names = TRUE,
  backtransform.mu = NULL,
  backtransform.sigma = NULL,
  backtransform.k = NULL,
  backtransform.rho = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{effects}{[character] Should the CIs/p-values for all coefficients be output (\code{"all"}),
or only for mean coefficients (\code{"mean"}),
or only for variance coefficients (\code{"variance"}),
or only for correlation coefficients (\code{"correlation"}).}

\item{level}{[numeric,0-1] the confidence level of the confidence intervals.}

\item{null}{[numeric vector] the value of the null hypothesis relative to each coefficient.}

\item{type.object}{[character] Set this argument to \code{"gls"} to obtain the output from the gls object and related methods.}

\item{strata}{[character vector] When not \code{NULL}, only output coefficient relative to specific levels of the variable used to stratify the mean and covariance structure.}

\item{df}{[logical] Should a Student's t-distribution be used to model the distribution of the coefficient. Otherwise a normal distribution is used.}

\item{type.information, transform.sigma, transform.k, transform.rho, transform.names}{are passed to the \code{vcov} method.}

\item{backtransform.mu, backtransform.sigma, backtransform.k, backtransform.rho}{[function] possible backtransformation for the estimate, lower and upper bounds of the confidence interval.}

\item{...}{Not used. For compatibility with the generic method.}

\item{param}{Not used. For compatibility with the generic method.}
}
\value{
A data.frame containing for each coefficient (in rows): \itemize{
\item column estimate: the estimate.
\item column se: the standard error.
\item column statistic: the test statistic.
\item column df: the degree of freedom.
\item column lower: the lower bound of the confidence interval.
\item column upper: the upper bound of the confidence interval.
\item column null: the null hypothesis.
\item column p.value: the p-value relative to the null hypothesis.
}
}
\description{
Compute confidence intervals (CIs) and p-values for the coefficients of a linear mixed model.
}
\seealso{
the function \code{multcomp::glht} to perform inference about linear combinations of coefficients and adjust for multiple comparisons.
}
