% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.resample}
\alias{confint.resample}
\title{Resampling Confidence Intervals for Linear Mixed Model}
\usage{
\method{confint}{resample}(
  object,
  parm = NULL,
  null = NULL,
  level = 0.95,
  method = NULL,
  columns = NULL,
  correction = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{reample} object.}

\item{parm}{Not used. For compatibility with the generic method.}

\item{null}{[numeric vector] the value of the null hypothesis relative to each coefficient.
Only relevant for when using bootstrap.}

\item{level}{[numeric,0-1] the confidence level of the confidence intervals.}

\item{method}{[character] method used to compute the confidence intervals and p-values: \code{"percentile"}, \code{"gaussian"}, or \code{"studentized"}.}

\item{columns}{[character vector] Columns to be output.
Can be any of \code{"estimate"}, \code{"sample.estimate"}, \code{"se"}, \code{"sample.se"}, \code{"null"}, \code{"lower"}, \code{"upper"}, \code{"p.value"}.}

\item{correction}{[logical] correction to ensure non-0 p-values when using the percentile method,
e.g. with permutations the p.value is evaluated as (#more extreme + 1)/(n.sample + 1) instead of (#more extreme)/(n.sample).}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Compute confidence intervals for linear mixed model using resampling (permutation or bootstrap).
}
\details{
Argument \bold{correction}: if we denote by \code{n.sample} the number of permutations that have been performed,
having the correction avoids p-values of 0 by ensuring they are at least \code{as.numeric(correction)/(n.sample+as.numeric(correction))},
e.g. at least 0.000999001 for a thousand permutations. \cr \cr

Argument \bold{correction} (bootstrap): percentile confidence intervals are computed based on the quantile of the resampling distribution. \cr
Gaussian confidence intervals and p-values are computed by assuming a normal distribution and estimating its variance based on the bootstrap distribution. \cr
Studentized confidence intervals are computed using quantiles based on the boostrap distribution after it has been centered and rescaled by the point estimate and its standard error
Percentile and Studentized p-values are computed by finding the coverage level such that one of the bound of the confidence interval equals the null. \cr \cr

Argument \bold{correction} (bootstrap): 
Percentile p-values are computed based on the proportion of times the estimate was more extreme than the permutation estimates. \cr
Gaussian p-values are computed by assuming a normal distribution and estimating its variance based on the permutation distribution. \cr
Studentized p-values are computed based on the proportion of times the test statistic was more extreme than the permutation test statistic. \cr
No confidence intervals are provided.
}
