% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Compute correlation matrix}
\usage{
correlate(
  formula,
  data,
  repetition,
  use = "everything",
  method = "pearson",
  collapse.value = ".",
  collapse.var = ".",
  na.rm
)
}
\arguments{
\item{formula}{[formula] on the left hand side the outcome(s) and on the right hand side the grouping variables.
E.g. Y1+Y2 ~ Gender will compute for each gender the correlation matrix for Y1 and the correaltion matrix for Y2.}

\item{data}{[data.frame] dataset containing the observations.}

\item{repetition}{[formula] Specify the structure of the data: the time/repetition variable and the grouping variable, e.g. ~ time|id.}

\item{use}{[character] method for computing correlation in the presence of missing values: \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"}, \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.
Passed to \code{\link{cor}}.}

\item{method}{[character] type of correlation coefficient: \code{"pearson"}, \code{"kendall"}, \code{"spearman"}.
Passed to \code{\link{cor}}.}

\item{collapse.value}{[character] symbol used to combine covariate values when using multiple grouping variables.}

\item{collapse.var}{[character] symbol used to combine variable names to be pasted left of the covariate values when using multiple grouping variables.
Can be disabled setting it to \code{NULL} or \code{FALSE}.}

\item{na.rm}{[logical] not used. The user may expect this argument though so it is added to help the user with a message pointing toward the argument \code{use}.}
}
\description{
Compute correlation matrix for multiple variables and/or multiple groups.
}
\examples{
#### simulate data (wide format) ####
data(gastricbypassL, package = "LMMstar")

## compute correlations (composite time variable)
e.S <- correlate(weight ~ 1, data = gastricbypassL, repetition = ~time|id)
e.S
as.matrix(e.S)

e.S21 <- correlate(glucagonAUC + weight ~ 1, data = gastricbypassL,
                  repetition = ~time|id, use = "pairwise")
e.S21
as.array(e.S21)
as.matrix(e.S21, index = "weight")

gastricbypassL$sex <- as.numeric(gastricbypassL$id) \%\% 2
e.S12 <- correlate(weight ~ sex, data = gastricbypassL,
                   repetition = ~time|id, use = "pairwise")
e.S12
as.array(e.S12)

e.S22 <- correlate(glucagonAUC + weight ~ sex, data = gastricbypassL,
                   repetition = ~time|id, use = "pairwise")
e.S22
as.array(e.S22)
}
\seealso{
\code{\link{as.array}} to convert the output to an array or \code{\link{as.matrix}} to convert the output to a matrix (when a single outcome and no grouping variable). \cr
\code{\link{summarize}} for other summary statistics (e.g. mean, standard deviation, ...).
}
