% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{blandAltmanW}
\alias{blandAltmanW}
\title{Data From The Bland Altman Study (Wide Format)}
\usage{
data(blandAltmanW)
}
\description{
Data From The Bland Altman Study where two methods to measure the peak expiratory flow rate (PEFR) where compared.
This dataset is in the wide format (i.e. one line per patient).

\itemize{
\item \code{id}: patient identifier.
\item \code{wright1}: first measurement made with a Wright peak flow meter.
\item \code{wright2}: second measurement made with a Wright peak flow meter.
\item \code{mini1}: first measurement made with a mini Wright peak flow meter.
\item \code{mini2}: second measurement made with a mini Wright peak flow meter.
}
}
\references{
Bland & Altman, Statistical methods for assessing agreement between two methods of clinical measurement, Lancet, 1986; i: 307-310.
}
\keyword{data}
