% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{bloodpressureL}
\alias{bloodpressureL}
\title{Data From The Blood Pressure Study (Long Format)}
\usage{
data(bloodpressureL)
}
\description{
Data from a cross-over trial comparing the impact of three formulations of a drug on the blood pressure.
The study was conducted on 12 male volunteers randomly divided into tree groups
and receiving each of the three formulations with a wash-out period of one week.

\itemize{
\item id Patient identifier
\item sequence sequence of treatment 
\item treatment Formulation of the treatment:
A (50 mg tablet)
B (100 mg tablet)
C (sustained-release formulation capsule)
\item period time period (in weeks)
\item duration duration of the drug (in hours)
}
}
\references{
TO ADD
}
\keyword{data}
