% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{calciumL}
\alias{calciumL}
\title{Data From The Calcium Supplements Study (Long Format)}
\usage{
data(calciumL)
}
\description{
Data from a randomized study including 112 girls at age 11 investigate the effect of a calcium supplement (n=55) vs. placebo (n=57)
on bone mineral density over a 2 year follow-up. The clinical question is: does a calcium supplement help to increase bone gain in adolescent women?
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item \code{girl}: patient identifier.
\item \code{grp}: treatment group: calcium supplement (coded \code{C}) or placebo (coded \code{P}).
\item \code{visit}: visit index.
\item \code{bmd}: bone mineral density (mg/cm3).
\item \code{time.obs}: visit time (in years).
\item \code{time.num}: scheduled visit time (numeric variable, in years).
\item \code{time.fac}: scheduled visit time (factor variable).
}
}
\references{
TO ADD
}
\keyword{data}
