% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{calciumW}
\alias{calciumW}
\title{Data From The Calcium Supplements Study (Wide Format)}
\usage{
data(calciumW)
}
\description{
Data from a randomized study including 112 girls at age 11 investigate the effect of a calcium supplement (n=55) vs. placebo (n=57)
on bone mineral density over a 2 year follow-up. The clinical question is: does a calcium supplement help to increase bone gain in adolescent women?
This dataset is in the wide format (i.e. one line per patient).

\itemize{
\item \code{girl}: patient identifier
\item \code{grp}: treatment group: calcium supplement (coded \code{C}) or placebo (coded \code{P}).
\item \code{obstime1}: time after the start of the study at which the first visit took place (in years).
\item \code{obstime2}: time after the start of the study at which the second visit took place (in years).
\item \code{obstime3}: time after the start of the study at which the third visit took place (in years).
\item \code{obstime4}: time after the start of the study at which the fourth visit took place (in years).
\item \code{obstime5}: time after the start of the study at which the fifth visit took place (in years).
\item \code{bmd1}: bone mineral density measured at the first visit (in mg/cm3).
\item \code{bmd2}: bone mineral density measured at the second visit (in mg/cm3).
\item \code{bmd3}: bone mineral density measured at the third visit (in mg/cm3).
\item \code{bmd4}: bone mineral density measured at the fourth visit (in mg/cm3).
\item \code{bmd5}: bone mineral density measured at the fifth visit (in mg/cm3).
}
}
\references{
Vonesh and Chinchilli 1997. Linear and Nonlinear models for the analysis of repeated measurement (Table 5.4.1 on page 228). New York: Marcel Dekker.
}
\keyword{datasets}
