% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ckdL}
\alias{ckdL}
\title{CKD long}
\usage{
data(ckdL)
}
\description{
TODO

\itemize{
\item \code{id}: patient identifier.
\item \code{allocation}:
\item \code{sex}:
\item \code{age}:
\item \code{visit}:
\item \code{time}:
\item \code{pwv}:
\item \code{aix}:
\item \code{dropout}:
}
}
\references{
TO ADD
}
\keyword{datasets}
