% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ckdW}
\alias{ckdW}
\title{CKD wide}
\usage{
data(ckdW)
}
\description{
TODO

\itemize{
\item id Patient identifier
\item allocation
\item sex
\item age
\item pwv0
\item pwv12
\item pwv24
\item aix0
\item aix12
\item aix24
\item dropout
}
}
\references{
TO ADD
}
\keyword{data}
