% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{gastricbypassL}
\alias{gastricbypassL}
\title{Data From The Gastric Bypass Study (Long Format)}
\usage{
data(gastricbypassL)
}
\description{
Data from the gastric bypass study
where the bodyweight and serum glucagon (a gut hormone) were measured in 20 obese subjects prior and after gastric bypass surgery.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item \code{id}: patient identifier.
\item \code{visit}: the visit index.
\item \code{time}: the time at which the visit took place.
\item \code{weight}: bodyweight (in kg) measured during the visit.
\item \code{glucagonAUC}: glucagon measured during the visit.
}
}
\references{
The effect of Roux-en-Y gastric bypass surgery on the gut mucosal gene expression profile and circulating gut hormones. \url{https://easddistribute.m-anage.com/from.storage?image=4iBH9mRQm1kfeEHULC2CxovdlyCtA1EHeVDdoffnZrAUGG9SHTO-U4ItnLU078eVkF1ZUZgYTy7THlTW3KSgFA2}
}
\keyword{data}
