% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ncgsW}
\alias{ncgsW}
\title{Data From National Cooperative Gallstone Study (Wide Format)}
\usage{
data(ncgsW)
}
\description{
Data from the National Cooperative Gallstone Study (NCGS),
a randomized study where the level of serum cholesterol was measured at baseline and after intake of high-dose chenondiol (750mg/day) or placebo.
This dataset is in the wide format (i.e. one line per patient).

\itemize{
\item \code{group}: treatment group (highdose or placebo).
\item \code{id}: patient identifier.
\item \code{cholest1}: cholesterol measurement at baseline (before treatment).
\item \code{cholest2}: cholesterol measurement at 6 months (after treatment).
\item \code{cholest3}: cholesterol measurement at 12 months (after treatment).
\item \code{cholest4}: cholesterol measurement at 20 months (after treatment).
\item \code{cholest5}: cholesterol measurement at 24 months (after treatment).
}
}
\references{
Grundy SM, Lan SP, Lachin J. The effects of chenodiol on biliary lipids and their association with gallstone dissolution in the National Cooperative Gallstone Study (NCGS). J Clin Invest. 1984 Apr;73(4):1156-66. doi: 10.1172/JCI111301.
}
\keyword{data}
