% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{onycholysisW}
\alias{onycholysisW}
\title{Data From The toenail onycholysis Study (Wide Format)}
\usage{
data(onycholysisW)
}
\description{
Data from the toenail onychomycosis study,
a randomized double-blind comparative study comparing Terbinafine (250 mg/day) against Itraconazole (200 mg/day) over 7 weeks.
This dataset is in the wide format (i.e. one line per patient).

\itemize{
\item \code{id}: patient identifier.
\item \code{group}: treatment arm to which the patient has been randomized.
\item \code{response1},...,\code{response7}: degree of onycholysis (separation of the nail plate from the nail-bed). Can be 0 for none or mild, 1 for moderate or severe.
\item \code{time1},...,\code{time7}: time elapsed between baseline and the measurement (in months).
}
}
\references{
De Backer et al. Twelve weeks of continuous oral therapy for toenail onychomycosis caused by dermatophytes: A double-blind comparative trial of terbinafine 250 mg/day versus itraconazole 200 mg/day,  Journal of the American Academy of Dermatology, (2020) 110. doi: 10.1016/s0190-9622(98)70486-4
}
\keyword{datasets}
