% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{potassiumSingleL}
\alias{potassiumSingleL}
\title{Data From The Potassium Intake Study (Long Format)}
\usage{
data(potassiumSingleL)
}
\description{
Data from the potassium intake study,
a randomized placebo-controlled crossover study where the effect of potassium supplement (90 mmol/day) on the renin-angiostensin-aldosteron system (RAAS) was assessed.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item id Patient identifier
\item sequence Treatment group to which the patient has been randomized.
\item period Time period.
\item treatment Treatment during the time period
\item auc Area under the curve of ?? during the time period
\item bsauc ??
\item aldo ??
}
}
\references{
Dreier et al. Effect of increased potassium intake on the reninangiotensinaldosterone system and subcutaneous resistance arteries: a randomized crossover study,
Nephrol Dial Transplant (2020) 110. doi: 10.1093/ndt/gfaa114
}
\keyword{data}
