% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{schoolL}
\alias{schoolL}
\title{Simulated Data with 3-level struture (Long Format)}
\usage{
data(schoolL)
}
\description{
Simulated data a nested structure: Student/Class/School and one outcome.

\itemize{
\item \code{school}:
\item \code{class}: 
\item \code{student}:
\item \code{outcome}:  
}
}
\keyword{data}
