% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sleepL}
\alias{sleepL}
\title{Data from a sleep study}
\usage{
data(sleepL)
}
\description{
\itemize{
\item \code{id}: patient identifier.
\item \code{deprivation}: 
\item \code{condition}: 
\item \code{vigilance}: 
\item \code{signal.34} and \code{signal.98}: outcome.
}
}
\keyword{datasets}
