% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{swabsL}
\alias{swabsL}
\title{Data From The SWABS Study (Long Format)}
\usage{
data(swabsL)
}
\description{
Data from the swabs study,
where the pneumococcus was studied in 18 families with different space available for the household.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item \code{crowding}: space available in the household.
\item \code{family}: family serial number
\item \code{name}: type of family member.
\item \code{swabs}: number of times the swab measurement was positive.
}
}
\references{
TODO
}
\keyword{datasets}
