% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{swabsW}
\alias{swabsW}
\title{Data From The SWABS Study (Wide Format)}
\usage{
data(swabsW)
}
\description{
Data from the swabs study,
where the pneumococcus was studied in 18 families with different space available for the household.
This dataset is in the wide format (i.e. one line per patient).

\itemize{
\item \code{crowding}: space available in the household.
\item \code{family}: family serial number
\item \code{mother}: number of times the swab measurement was positive for the mother.
\item \code{father}: number of times the swab measurement was positive for the father.
\item \code{child1}: number of times the swab measurement was positive for the first child.
\item \code{child2}: number of times the swab measurement was positive for the second child.
\item \code{child3}: number of times the swab measurement was positive for the third child.
}
}
\references{
Grundy SM, Lan SP, Lachin J. The effects of chenodiol on biliary lipids and their association with gallstone dissolution in the National Cooperative Gallstone Study (SWABS). J Clin Invest. 1984 Apr;73(4):1156-66. doi: 10.1172/JCI111301.
}
\keyword{datasets}
