% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{vasscoresL}
\alias{vasscoresL}
\title{Data From The VAS Study (Long Format)}
\usage{
data(vasscoresL)
}
\description{
Data from the VAS Study,
a randomized controlled clinial trial assessing the healing effect of topical zink sulfate on epidermal wound.
The study includes 30 heatlhy volunteers with induced wounds on each buttock which where subsequently treated with a different treatment for each wound.
Then the VAS-score (pain sensation on a 0-100mm visual analogue scale) was assessed after each treatment application and summarized by area under the curve.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item \code{id}: patient identifier.
\item \code{group}: treatment group to which the patient has been randomized.
\item \code{treat.num}:
\item \code{vas}: VAS-score relative to the wound.
\item \code{treatment}: Treatment used on the wound.
A: active treatment (zink shower gel),
B: placebo treatment (shower gel without zink),
C: control treatment (demineralized water).
}
}
\references{
TODO
}
\keyword{datasets}
