% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{vasscoresW}
\alias{vasscoresW}
\title{Data From The VAS Study (Wide Format)}
\usage{
data(vasscoresW)
}
\description{
Data from the VAS Study,
a randomized controlled clinial trial assessing the healing effect of topical zink sulfate on epidermal wound.
The study includes 30 heatlhy volunteers with induced wounds on each buttock which where subsequently treated with a different treatment for each wound.
Then the VAS-score (pain sensation on a 0-100mm visual analogue scale) was assessed after each treatment application and summarized by area under the curve.
This dataset is in the wide format (i.e. one line per patient).

\itemize{
\item \code{id}: patient identifier.
\item \code{group}: treatment group to which the patient has been randomized.
\item \code{vasA}: VAS-score when using a zink shower gel.
\item \code{vasB}: VAS-score when using a placebo treatment (shower gel without zink).
\item \code{vasC}: VAS-score when using a control treatment with demineralized water.
}
}
\references{
TODO
}
\keyword{data}
