% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df.residual.mlmm}
\alias{df.residual.mlmm}
\title{Residuals Degrees-of-Freedom From From Multiple Linear Mixed Model.}
\usage{
\method{df.residual}{mlmm}(object, ...)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{...}{Passed to \code{residuals.lmm}.}
}
\value{
A numeric vector with one element for each model.
}
\description{
Combine the residuals degrees-of-freedom from group-specific linear mixed models.
}
\details{
The residual degrees-of-freedom is estimated separately for each model using the sum of squared normalized residuals.
}
\keyword{methods}
