% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-calc_d2Omega.R
\name{.calc_d2Omega}
\alias{.calc_d2Omega}
\title{Second Derivative of the Residual Variance-Covariance Matrix}
\usage{
.calc_d2Omega(object, param, Omega, dOmega, Jacobian, dJacobian)
}
\arguments{
\item{param}{[named numeric vector] values of the parameters.}

\item{Omega}{[list of matrices] Residual Variance-Covariance Matrix for each pattern.}

\item{dOmega}{[list of matrices] First derivative of the residual Variance-Covariance Matrix for each pattern.}

\item{Jacobian}{[matrix] Jacobian of the reparametrisation.}

\item{dJacobian}{[array] First derivative of the Jacobian of the reparametrisation.}

\item{structure}{[structure]}
}
\description{
Second derivative of the residual variance-covariance matrix for given parameter values.
}
\examples{
\dontrun{
data(gastricbypassL, package = "LMMstar")
gastricbypassL$gender <- c("M","F")[as.numeric(gastricbypassL$id) \%\% 2+1]
dd <- gastricbypassL[!duplicated(gastricbypassL[,c("time","gender")]),]

## independence
Sid1 <- skeleton(IND(~1, var.time = "time"), data = dd)
Sid4 <- skeleton(IND(~1|id, var.time = "time"), data = dd)
Sdiag1 <- skeleton(IND(~visit), data = dd)
Sdiag4 <- skeleton(IND(~visit|id), data = dd)
Sdiag24 <- skeleton(IND(~visit+gender|id, var.time = "time"), data = dd)
param24 <- setNames(c(1,2,2,3,3,3,5,3),Sdiag24$param$name)

.calc_d2Omega(Sid4, param = c(sigma = 2))
.calc_d2Omega(Sdiag1, param = c(sigma = 1, k.visit2 = 2, k.visit3 = 3, k.visit4 = 4))
.calc_d2Omega(Sdiag4, param = c(sigma = 1, k.visit2 = 2, k.visit3 = 3, k.visit4 = 4))
.calc_d2Omega(Sdiag24, param = param24)

## compound symmetry
Scs4 <- skeleton(CS(~1|id, var.time = "time"), data = gastricbypassL)
Scs24 <- skeleton(CS(gender~time|id), data = gastricbypassL)

.calc_d2Omega(Scs4, param = c(sigma = 1,rho=0.5))
.calc_d2Omega(Scs4, param = c(sigma = 2,rho=0.5))
.calc_d2Omega(Scs24, param = c("sigma:F" = 2, "sigma:M" = 1,
                            "rho:F"=0.5, "rho:M"=0.25))

## unstructured
Sun4 <- skeleton(UN(~visit|id), data = gastricbypassL)
param4 <- setNames(c(1,1.1,1.2,1.3,0.5,0.45,0.55,0.7,0.1,0.2),Sun4$param$name)
Sun24 <- skeleton(UN(gender~visit|id), data = gastricbypassL)
param24 <- setNames(c(param4,param4*1.1),Sun24$param$name)

.calc_d2Omega(Sun4, param = param4)
.calc_d2Omega(Sun24, param = param24)
}
}
\keyword{internal}
