% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{.estimate}
\alias{.estimate}
\title{Optimizer for mixed models}
\usage{
.estimate(
  design,
  time,
  method.fit,
  type.information,
  transform.sigma,
  transform.k,
  transform.rho,
  precompute.moments,
  optimizer,
  init,
  n.iter,
  tol.score,
  tol.param,
  trace
)
}
\arguments{
\item{design}{[list] information about the mean and variance structure. Obtained using \code{.model.matrix.lmm}.}

\item{time}{[list] information about the time variable (e.g. unique values)}

\item{method.fit}{[character] Should Restricted Maximum Likelihoood (\code{"REML"}) or Maximum Likelihoood (\code{"ML"}) be used to estimate the model parameters?}

\item{type.information}{[character] Should the expected information be computed  (i.e. minus the expected second derivative) or the observed inforamtion (i.e. minus the second derivative).}

\item{transform.sigma, transform.k, transform.rho}{possible transformations for the variance parameters.}

\item{precompute.moments}{[logical] have certain key quantities be pre-computed (e.g. \eqn{X'X}).}

\item{init}{[numeric vector] values used to initialize the mean and parameters.}

\item{n.iter}{[integer,>0] maximum number of iterations.}

\item{tol.score}{[double,>0] convergence is not reached unless each element of the score is smaller (in absolute value) than this value.}

\item{tol.param}{[double,>0] convergence is not reached unless the change in parameter values between two iterations is smaller (in absolute value) than this value.}

\item{trace}{[1, 2, or 3] should each iteration be displayed?}
}
\description{
Optimization procedure for mixed model (REML or ML).
Alternate between one step of gradient descent to update the variance parameters
and a GLS estimation of the mean parameters.
}
\examples{
\dontrun{
## simulate data in the long format
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")

## fit Linear Mixed Model
LMMstar.options(optimizer = "gls")
eUN.gls <- lmm(Y ~ X1 + X2 + X5, repetition = ~visit|id, structure = "UN", data = dL, df = FALSE)

LMMstar.options(optimizer = "FS")
eUN.lmm <- lmm(Y ~ X1 + X2 + X5, repetition = ~visit|id, structure = "UN", data = dL, df = FALSE)

}

}
\keyword{internal}
