% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{.formulaStructure}
\alias{.formulaStructure}
\title{Extract Variable From Formula For VCOV Structure}
\usage{
.formulaStructure(formula, missing.time.ok = FALSE, missing.id.ok = FALSE)
}
\arguments{
\item{formula}{A formula or a list of two formulas.}

\item{missing.time.ok}{[logical] If FALSE an error is triggered when the function could not identify the time variable.}

\item{missing.id.ok}{[logical] If FALSE an error is triggered when the function could not identify the cluster variable.}
}
\description{
Extract the variables from the variance and correlation formula to be used to initialize the variance-covariance structure.
}
\examples{
\dontrun{
.formulaStructure(strata ~ time|id)
.formulaStructure( ~ time|id)
.formulaStructure(list( ~ gender+time|id,  ~ time|id))
.formulaStructure(strata ~ 1|id, missing.time.ok = TRUE)
.formulaStructure(strata ~ time, missing.id.ok = TRUE)
.formulaStructure( ~ time|id)
}
}
\keyword{internal}
