% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-initialization.R
\name{.initialize}
\alias{.initialize}
\title{Initialize Variance-Covariance Structure}
\usage{
.initialize(object, residuals, p, ssc)
}
\arguments{
\item{residuals}{[vector] vector of residuals.}

\item{structure}{[structure]}
}
\description{
Initialize the parameters of the variance-covariance structure using residual variance and correlations.
}
\examples{
\dontrun{
data(gastricbypassW, package = "LMMstar")
data(gastricbypassL, package = "LMMstar")
gastricbypassL$gender <- c("M","F")[as.numeric(gastricbypassL$id) \%\% 2+1]
dd <- gastricbypassL[!duplicated(gastricbypassL[,c("time","gender")]),]

eDD.lm <- lm(weight ~ visit, data = dd)
eGas.lm <- lm(weight ~ visit*gender, data = gastricbypassL)

## independence
Sid1 <- .skeleton(IND(~1, var.time = "time"), data = dd)
Sid4 <- .skeleton(IND(~1|id, var.time = "time"), data = dd)
Sdiag1 <- .skeleton(IND(~visit), data = dd)
Sdiag4 <- .skeleton(IND(~visit|id), data = dd)
Sdiag24 <- .skeleton(IND(~visit+gender|id, var.time = "time"), data = gastricbypassL)

.initialize(Sid1, residuals = residuals(eDD.lm))
## sd(residuals(eDD.lm))
.initialize(Sdiag4, residuals = residuals(eDD.lm))
## tapply(residuals(eDD.lm),dd$visit,sd)
.initialize(Sdiag24, residuals = residuals(eGas.lm))
## tapply(residuals(eGas.lm),interaction(gastricbypassL[,c("visit","gender")]),sd)

## compound symmetry
Scs4 <- .skeleton(CS(~1|id, var.time = "time"), data = gastricbypassL)
Scs24 <- .skeleton(CS(gender~time|id), data = gastricbypassL)

.initialize(Scs4, residuals = residuals(eGas.lm))
## cor(gastricbypassW[,c("weight1","weight2","weight3","weight4")])
.initialize(Scs24, residuals = residuals(eGas.lm))

## unstructured
Sun4 <- .skeleton(UN(~visit|id), data = gastricbypassL)
Sun24 <- .skeleton(UN(gender~visit|id), data = gastricbypassL)

.initialize(Sun4, residuals = residuals(eGas.lm))
.initialize(Sun24, residuals = residuals(eGas.lm))
}
}
\keyword{internal}
