% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-skeleton.R
\name{.skeleton}
\alias{.skeleton}
\title{Parametrization of Covariance Structure}
\usage{
.skeleton(structure, data)
}
\arguments{
\item{structure}{[structure]}

\item{data}{[data.frame] dataset}
}
\description{
Parametrization of Covariance Structure
}
\examples{
\dontrun{
data(gastricbypassL, package = "LMMstar")
gastricbypassL$gender <- c("M","F")[as.numeric(gastricbypassL$id) \%\% 2+1]
dd <- gastricbypassL[!duplicated(gastricbypassL[,c("time","gender")]),]

## independence
.skeleton(IND(~1, var.time = "time"), data = dd)
.skeleton(IND(~time), data = dd)
.skeleton(IND(~time|id), data = dd)
.skeleton(IND(~time+gender|id, var.time = "time"), data = dd)
.skeleton(IND(gender~time|id, var.time = "time"), data = dd)

## compound symmetry
.skeleton(CS(~1|id, var.time = "time"), data = gastricbypassL)
.skeleton(CS(~time|id), data = gastricbypassL)
.skeleton(CS(gender~time|id), data = gastricbypassL)

## unstructured
.skeleton(UN(~visit|id), data = gastricbypassL)
.skeleton(UN(~time|id), data = gastricbypassL)
.skeleton(UN(~visit+gender|id, var.time = "time"), data = gastricbypassL)
.skeleton(UN(gender~visit|id), data = gastricbypassL)
}
}
\keyword{internal}
