% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.coef.R
\name{dummy.coef.lmm}
\alias{dummy.coef.lmm}
\title{Extract Mean Coefficients in Original Coding from a Linear Mixed Model}
\usage{
\method{dummy.coef}{lmm}(object, use.na = FALSE, ...)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{use.na}{[logical] Should \code{NA} or \code{0} be used to represent excluded coefficients in singular models.}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
This expands the mean coefficients of the linear mixed model into one coefficient per level of the original variable,
i.e., including the reference level(s) where the fitted coefficients are 0.
}
\examples{
## simulate data in the long format
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")
dL$group <- as.factor(paste0("d",dL$X1))

## fit mixed model with interaction
eUN.lmm <- lmm(Y ~ visit*group, repetition =~visit|id, data = dL, df = FALSE)
dummy.coef(eUN.lmm)

## fit mixed model with baseline constraint
dL$drug <- dL$group
dL[dL$visit==1,"drug"] <- "d0"
eUN.clmm <- lmm(Y ~ visit + visit:drug, repetition =~visit|id, data = dL, df = FALSE)
dummy.coef(eUN.clmm)

dL$drug <- factor(dL$group, levels = c("none","d0","d1"))
dL[dL$visit==1,"drug"] <- "none"
eUN.clmm.none <- lmm(Y ~ visit:drug, repetition =~visit|id, data = dL, df = FALSE)
dummy.coef(eUN.clmm.none)

}
