% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effects.lmm}
\alias{effects.lmm}
\title{Effects Derived For Linear Mixed Model}
\usage{
\method{effects}{lmm}(
  object,
  type,
  variable,
  conditional = NULL,
  rhs = NULL,
  repetition = NULL,
  multivariate = FALSE,
  prefix.time = "t=",
  prefix.var = TRUE,
  sep.var = ",",
  ...
)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{type}{[character] should all mean coefficients with respect to a variable be tested (\code{"mean"})
or the average counterfactual outcome (\code{"aoc"}) or average treatment effect (\code{"ate"}) be estimated?}

\item{variable}{[character] the variable relative to which the effect should be computed.}

\item{conditional}{[character] variable conditional to which the average conterfactual outcome or treatment effect should be computed.}

\item{rhs}{[numeric] the right hand side of the hypothesis.}

\item{repetition}{[character vector] repetition at which the effect should be assessed. By default it will be assessed at all repetitions.}

\item{multivariate}{[logical] should a multivariate Wald test be used to simultaneously test all null hypotheses.}

\item{prefix.time}{[character] When naming the estimates, text to be pasted before the value of the repetition variable.
Only relevant when \code{type = "aoc"} or \code{type = "ate"}.}

\item{prefix.var}{[logical] When naming the estimates, should the variable name be added or only the value?}

\item{sep.var}{[character] When naming the estimates, text to be pasted between the values to condition on.
Only relevant when \code{type = "aoc"} or \code{type = "ate"}.}

\item{...}{Arguments passed to \code{\link{anova.lmm}}.}
}
\description{
Estimate mean effects from linear mixed models relative to a variable.
}
\examples{
#### simulate data in the long format ####
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")

#### fit Linear Mixed Model ####
eUN.lmm <- lmm(Y ~ visit + X1 + X2 + X5,
               repetition = ~visit|id, structure = "UN", data = dL)

effects(eUN.lmm, type = "mean", variable = "X1")
effects(eUN.lmm, type = "aco", variable = "X1")
effects(eUN.lmm, type = "ate", variable = "X1")
effects(eUN.lmm, type = "ate", variable = "X1", repetition = "3")

#### fit Linear Mixed Model with interaction ####
dL$X1.factor <- as.factor(dL$X1)
dL$X2.factor <- as.factor(dL$X2)
eUN.lmmI <- lmm(Y ~ visit + X1.factor * X2.factor + X5,
               repetition = ~visit|id, structure = "UN", data = dL)

effects(eUN.lmmI, type = "aco", variable = "X1.factor", conditional = c("X2.factor"), repetition = "3")
effects(eUN.lmmI, type = "ate", variable = "X1.factor", conditional = c("X2.factor"), repetition = "3")
effects(eUN.lmmI, type = "ate", variable = "X1.factor", conditional = list(X5=0:1), repetition = "3")
effects(eUN.lmmI, type = "ate", variable = "X1.factor", conditional = list(X5=0:5), repetition = "3")
}
