% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.mlmm}
\alias{estimate.mlmm}
\title{Delta Method for Multiple Linear Mixed Models}
\usage{
\method{estimate}{mlmm}(
  x,
  f,
  df = FALSE,
  robust = FALSE,
  type.information = NULL,
  level = 0.95,
  method.numDeriv = NULL,
  average = FALSE,
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{mlmm} object.}

\item{f}{[function] function taking as input \code{p}, a vector containing the group-specific linear mixed model parameters, which outputs the parameter(s) of interest.
Can accept extra-arguments.}

\item{df}{[logical] Should degree-of-freedom, computed using Satterthwaite approximation, for the parameter of interest be output.
Can also be a numeric vector providing providing the degrees-of-freedom relative to each estimate.}

\item{robust}{[logical] Should robust standard errors (aka sandwich estimator) be output instead of the model-based standard errors. 
Can also be \code{2} compute the degrees-of-freedom w.r.t. robust standard errors instead of w.r.t. model-based standard errors.}

\item{type.information}{[character] Should the expected information be used  (i.e. minus the expected second derivative) or the observed inforamtion (i.e. minus the second derivative).}

\item{level}{[numeric,0-1] the confidence level of the confidence intervals.}

\item{method.numDeriv}{[character] method used to approximate the gradient: either \code{"simple"} or \code{"Richardson"}.
Passed to \code{numDeriv::jacobian}.}

\item{average}{[logical] is the estimand the average output of argument \code{f}?
Otherwise consider each individual output of argument \code{f}.}

\item{transform.sigma}{[character] Transformation used on the variance coefficient for the reference level. One of \code{"none"}, \code{"log"}, \code{"square"}, \code{"logsquare"} - see details.}

\item{transform.k}{[character] Transformation used on the variance coefficients relative to the other levels. One of \code{"none"}, \code{"log"}, \code{"square"}, \code{"logsquare"}, \code{"sd"}, \code{"logsd"}, \code{"var"}, \code{"logvar"} - see details.}

\item{transform.rho}{[character] Transformation used on the correlation coefficients. One of \code{"none"}, \code{"atanh"}, \code{"cov"} - see details.}

\item{...}{extra arguments passed to \code{f}.}
}
\description{
Estimate standard errors, confidence intervals, and p-values for a smooth transformation of parameters from group-specific linear mixed models.
}
