% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.mlmm}
\alias{fitted.mlmm}
\title{Fitted Outcome Value by Multiple Linear Mixed Models.}
\usage{
\method{fitted}{mlmm}(object, newdata = NULL, simplify = TRUE, ...)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{newdata}{[data.frame] the covariate values for each cluster along with the \code{by} variable indicating which linear mixed model should be used.}

\item{simplify}{[logical] simplify the column names (no mention of the time variable) and remove the cluster, by, and time column when possible.}

\item{...}{additional argument passed the \code{\link{predict.lmm}}.}
}
\description{
Evaluate the expected mean conditional to covariates
or the expected missing outcome values conditional to observed outcome values and covariates.
based on group-specific linear mixed models.
}
