% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{gastricbypassW}
\alias{gastricbypassW}
\title{Data From The Gastric Bypass Study (Wide Format)}
\usage{
data(gastricbypassW)
}
\description{
Data from the gastric bypass study
where the bodyweight and serum glucagon (a gut hormone) were measured in 20 obese subjects prior and after gastric bypass surgery.
This dataset is in the wide format (i.e. one line per patient).

\itemize{
\item id Patient identifier
\item weight1 Bodyweight (in kg) 3 months before surgery.
\item weight2 Bodyweight (in kg) 1 week before surgery.
\item weight3 Bodyweight (in kg) 1 week after surgery.
\item weight4 Bodyweight (in kg) 3 months after surgery.
\item glucagonAUC1 Glucagon value 3 months before surgery.
\item glucagonAUC2 Glucagon value 1 week before surgery.
\item glucagonAUC3 Glucagon value 1 week after surgery.
\item glucagonAUC4 Glucagon value 3 months after surgery.
}
}
\references{
The effect of Roux-en-Y gastric bypass surgery on the gut mucosal gene expression profile and circulating gut hormones. \url{https://easddistribute.m-anage.com/from.storage?image=4iBH9mRQm1kfeEHULC2CxovdlyCtA1EHeVDdoffnZrAUGG9SHTO-U4ItnLU078eVkF1ZUZgYTy7THlTW3KSgFA2}
}
\keyword{data}
