% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{getVarCov.lmm}
\alias{getVarCov.lmm}
\title{Depreciated Extractor of the Residual Variance-Covariance Matrix}
\usage{
\method{getVarCov}{lmm}(obj, ...)
}
\arguments{
\item{obj}{a \code{lmm} object.}

\item{...}{other arguments.}
}
\value{
Nothing
}
\description{
Depreciated extractor of the residual variance-covariance matrix.
}
\seealso{
\code{\link{sigma.lmm}}
}
