% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{ID}
\alias{ID}
\title{identity Structure}
\usage{
ID(formula, var.cluster, var.time, add.time)
}
\arguments{
\item{formula}{formula indicating on which variable to stratify the residual variance (left hand side).}

\item{var.cluster}{[character] cluster variable.}

\item{var.time}{[character] time variable.}

\item{add.time}{not used.}
}
\value{
An object of class \code{ID} that can be passed to the argument \code{structure} of the \code{lmm} function.
}
\description{
Variance-covariance structure where the residuals are independent and identically distributed.
Can be stratified on a categorical variable.
}
\details{
A typical formula would be \code{~1}. It will model: \itemize{
\item \eqn{\sigma^2_{j}=\sigma^2}: a variance constant over repetitions.
\item \eqn{\rho_{j,j'}=0}: no correlation.
}
With 4 repetitions (\eqn{(j,j') \in \{1,2,3,4\}^2} and \eqn{j\neq j'}) this leads to the following residuals variance-covariance matrix:
\deqn{ \begin{bmatrix} \sigma^2 & 0 & 0 & 0 \\ 0 & \sigma^2 & 0 & 0 \\ 0 & 0 & \sigma^2 & 0 \\ 0 & 0 & 0 & \sigma^2 \end{bmatrix}}
It is possibly to stratify covariance pattern on a categorical variable (e.g. sex) using a formula such as \code{~sex}. This will lead to \eqn{\sigma^2_{j,sex}=\sigma_{sex}^2} and \eqn{\rho_{j,j',sex}=0}, i.e.:
\deqn{ \begin{bmatrix} \sigma_{\text{female}}^2 & 0 & 0 & 0 \\ 0 & \sigma_{\text{female}}^2 & 0 & 0 \\ 0 & 0 & \sigma_{\text{female}}^2 & 0 \\ 0 & 0 & 0 & \sigma_{\text{female}}^2 \end{bmatrix}
\text{ and }
      \begin{bmatrix} \sigma_{\text{male}}^2 & 0 & 0 & 0 \\ 0 & \sigma_{\text{male}}^2 & 0 & 0 \\ 0 & 0 & \sigma_{\text{male}}^2 & 0 \\ 0 & 0 & 0 & \sigma_{\text{male}}^2 \end{bmatrix}
}
}
\examples{
\dontrun{
data(gastricbypassL, package = "LMMstar")
gastricbypassL$sex <- factor(as.numeric(gastricbypassL$id) \%\% 2,
                             labels = c("female", "male"))

## default: single variance parameter
eID.default <- lmm(glucagonAUC ~ visit, repetition = ~visit|id,
                 structure = ID, data = gastricbypassL)
sigma(eID.default)
model.tables(eID.default, effects = "variance")

## stratified: one variance parameter per strata
eID.strata <- lmm(glucagonAUC ~ visit, repetition = ~visit|id,
                 structure = ID(sex~1), data = gastricbypassL)
sigma(eID.strata)
model.tables(eID.strata, effects = "variance")

## same as
eID.reg <- lmm(glucagonAUC ~ visit, repetition = ~visit|id,
                 structure = ID(~sex), data = gastricbypassL)
sigma(eID.reg)
model.tables(eID.reg, effects = "variance")
}
}
\keyword{multivariate}
