% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.R
\name{iid.mlmm}
\alias{iid.mlmm}
\title{Extract the Influence Function From Multiple Linear Mixed Models}
\usage{
\method{iid}{mlmm}(x, effects = "contrast", p = NULL, ordering = "by", simplify = TRUE, ...)
}
\description{
Extract the influence function for ML or REML estimators of parameters from group-specific linear mixed models.
}
