% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{IND}
\alias{IND}
\title{Independence Structure}
\usage{
IND(formula, var.cluster, var.time, add.time)
}
\arguments{
\item{formula}{formula indicating variables influencing the residual variance,
using either as a multiplicative factor (right hand side) or stratification (left hand side) to model their effect.}

\item{var.cluster}{[character] cluster variable.}

\item{var.time}{[character] time variable.}

\item{add.time}{Should the default formula (i.e. when \code{NULL}) contain a time effect.}
}
\value{
An object of class \code{IND} that can be passed to the argument \code{structure} of the \code{lmm} function.
}
\description{
Variance-covariance structure where the residuals are independent but may have different variance.
Can be stratified on a categorical variable.
}
\details{
A typical formula would be \code{~time}. It will model: \itemize{
\item \eqn{\sigma^2_{j}=\sigma^2 k^2_j}: a variance specific to each repetition. This is achieved using a baseline standard deviation parameter (\eqn{\sigma}) and a multiplier parameter (\eqn{k_j}) for each repetition \eqn{j}, with \eqn{k_1=1}.
\item \eqn{\rho_{j,j'}=0}: no correlation
}
With 4 repetitions (\eqn{(j,j') \in \{1,2,3,4\}^2} and \eqn{j\neq j'}) this leads to the following residuals variance-covariance matrix:
\deqn{ \begin{bmatrix} \sigma^2 & 0 & 0 & 0 \\ 0 & k_2^2\sigma^2 & 0 & 0 \\ 0 & 0 & k_3^2\sigma^2 & 0 \\ 0 & 0 & 0 & k^2_4\sigma^2 \end{bmatrix}}
It is possibly to stratify the covariance pattern on a categorical variable (e.g. sex) using a formula such as \code{sex~time}. This will lead to:
\deqn{ \begin{bmatrix} \sigma_{\text{female}}^2 & 0 & 0 & 0 \\ 0 & k_{2:\text{female}}^2\sigma_{\text{female}}^2 & 0 & 0 \\ 0 & 0 & k_{3:\text{female}}^2\sigma_{\text{female}}^2 & 0 \\ 0 & 0 & 0 & k^2_{4:\text{female}}\sigma_{\text{female}}^2 \end{bmatrix}
\text{and}
\begin{bmatrix} \sigma_{\text{male}}^2 & 0 & 0 & 0 \\ 0 & k_{2:\text{male}}^2\sigma_{\text{male}}^2 & 0 & 0 \\ 0 & 0 & k_{3:\text{male}}^2\sigma_{\text{male}}^2 & 0 \\ 0 & 0 & 0 & k^2_{4:\text{male}}\sigma_{\text{male}}^2 \end{bmatrix}
}
Instead of stratifying one can also make the multiplier parameter specific to arbitrary (categorical) variables, e.g.  \code{~sex}:
\deqn{ \begin{bmatrix} \sigma^2 & 0 & 0 & 0 \\ 0 & k_{\text{female}:2}^2\sigma^2 & 0 & 0 \\ 0 & 0 & k_{\text{female}:3}^2\sigma^2 & 0 \\ 0 & 0 & 0 & k^2_{\text{female}:4}\sigma^2 \end{bmatrix}
\text{and}
\begin{bmatrix} \sigma^2 & 0 & 0 & 0 \\ 0 & k_{\text{male}:2}^2\sigma^2 & 0 & 0 \\ 0 & 0 & k_{\text{male}:3}^2\sigma^2 & 0 \\ 0 & 0 & 0 & k^2_{\text{male}:4}\sigma^2 \end{bmatrix}
}
which is just a re-parametrisation the stratified structure since, by default, the variance is taken dependent of a time effect.
Removing the time effect by setting \code{add.time} to \code{FALSE}) leads to:
\deqn{ \begin{bmatrix} \sigma^2 & 0 & 0 & 0 \\ 0 & \sigma^2 & 0 & 0 \\ 0 & 0 & \sigma^2 & 0 \\ 0 & 0 & 0 & \sigma^2 \end{bmatrix}
\text{and}
\begin{bmatrix} \sigma^2 k_{\text{male}}^2 & 0 & 0 & 0 \\ 0 & \sigma^2 k_{\text{male}}^2 & 0 & 0 \\ 0 & 0 & \sigma^2 k_{\text{male}}^2 & 0 \\ 0 & 0 & 0 & \sigma^2 k_{\text{male}}^2 \end{bmatrix}
}
}
\examples{
\dontrun{
data(gastricbypassL, package = "LMMstar")
gastricbypassL$sex <- factor(as.numeric(gastricbypassL$id) \%\% 2,
                             labels = c("female", "male"))

## default: repetition specific variance
eIND.default <- lmm(glucagonAUC ~ visit, repetition = ~visit|id,
                 structure = IND, data = gastricbypassL)
sigma(eIND.default)
paramIND.default <- model.tables(eIND.default, effects = "variance")
paramIND.default
paramIND.default[1,"estimate"]^2 * c(1,paramIND.default[-1,"estimate"]^2)

## stratified: repetition and strata specific variance parameter
eIND.strata <- lmm(glucagonAUC ~ visit, repetition = ~visit|id,
                   structure = IND(sex~1), data = gastricbypassL)
sigma(eIND.strata)
model.tables(eIND.strata, effects = "variance")

## same as
eIND.reg <- lmm(glucagonAUC ~ visit, repetition = ~visit|id,
                 structure = IND(~sex), data = gastricbypassL)
sigma(eIND.reg) 
model.tables(eIND.reg, effects = "variance") 

## to not only use sex dependent variance
eIND.reg0 <- lmm(glucagonAUC ~ visit, repetition = ~visit|id,
                 structure = IND(~sex, add.time = FALSE), data = gastricbypassL)
sigma(eIND.reg0)
model.tables(eIND.reg0, effects = "variance") 
}

}
\keyword{multivariate}
