% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\name{levels.lmm}
\alias{levels.lmm}
\title{Contrasts and Reference Level}
\usage{
\method{levels}{lmm}(x)
}
\arguments{
\item{x}{an \code{lmm} object}
}
\value{
a list with two elements \itemize{
\item all: contrast matrix for each categorical or factor variable
\item reference: reference level: one value for each categorical variable
}
}
\description{
Contrasts and reference level used when modeling the mean in a linear mixed modek.
}
\keyword{methods}
