% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMstar-package.R
\docType{package}
\name{LMMstar-package}
\alias{LMMstar-package}
\title{LMMstar package: Helper functions for handling repeated measurements in R}
\description{
Companion R package for the course "Statistical analysis
of correlated and repeated measurements for health science
researchers" taught by the section of Biostatistics of the University
of Copenhagen. It provides functions for computing summary statistics
and obtainign graphical displays of longitudinal data, as well as for
statistical modeling and statistical inference using mixed models.
}
