% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{recover_data.lmm}
\alias{recover_data.lmm}
\alias{emm_basis.lmm}
\title{Link to emmeans package}
\usage{
\method{recover_data}{lmm}(object, ...)

\method{emm_basis}{lmm}(object, trms, xlev, grid, ...)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{...}{Not used. For compatibility with the generic method.}

\item{trms}{see \code{emmeans::emm_basis} documentation}

\item{xlev}{see \code{emmeans::emm_basis} documentation}

\item{grid}{see \code{emmeans::emm_basis} documentation}
}
\value{
dataset or list used by the emmeans package.
}
\description{
Link to emmeans package. Not meant for direct use.
}
