% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.mlmm}
\alias{logLik.mlmm}
\title{Log-Likelihood From Multiple Linear Mixed Models}
\usage{
\method{logLik}{mlmm}(object, newdata = NULL, p = NULL, indiv = FALSE, ...)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{newdata}{[data.frame] dataset relative to which the log-likelihood should be computed. Only relevant if differs from the dataset used to fit the model.}

\item{p}{[list of numeric vector] values for the model parameters at which to evaluate the log-likelihood.
Only relevant if differs from the fitted values.}

\item{indiv}{[logical] Should the contribution of each cluster to the log-likelihood be output? Otherwise output the sum of all clusters of the derivatives.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A numeric vector when \code{indiv=FALSE} (log-likelihood per model)
or a matrix of numeric values when when \code{indiv=TRUE} with one line for each cluster and one column for each model.
}
\description{
Extract or compute the log-likelihood from each group-specific Linear Mixed Model.
}
\details{
\bold{indiv}: only relevant when using maximum likelihood. Must be \code{FALSE} when using restricted maximum likelihood.
}
\keyword{methods}
