% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.R
\name{model.frame.lmm}
\alias{model.frame.lmm}
\title{Extracting the Model Frame from a Linear Mixed Model}
\usage{
\method{model.frame}{lmm}(formula, effects = "mean", unique = FALSE, ...)
}
\arguments{
\item{formula}{[lmm] linear mixed model object}

\item{unique}{[logical] should the dataset contains only unique sets of covariates instead of the ones relative to the observations}

\item{...}{Not used. For compatibility with the generic method.}

\item{effect}{[character vector] covariates to be consider: relative to the mean (\code{"mean"}), variance (\code{"variance"}), or correlation (\code{"correlation"}) structure.
Or all of the previous (\code{"all"}).}
}
\description{
Variables needed to fit the Linear Mixed Model.
}
\examples{
data("armd.wide", package = "nlmeU")
e.lmH <- lmm(visual52 ~ lesion, structure = IND(~treat.f), data = armd.wide)
model.frame(e.lmH, effects = "mean", unique = TRUE)
model.frame(e.lmH, effects = "all", unique = TRUE)
model.frame(e.lmH)

}
