% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.tables.R
\name{model.tables}
\alias{model.tables}
\alias{model.tables.lmm}
\title{Statistical Inference for Linear Mixed Model}
\usage{
\method{model.tables}{lmm}(x, ...)
}
\arguments{
\item{x}{a \code{lmm} object.}

\item{...}{arguments to be passed to the \code{confint} method. Should not contain the argument \code{column}.}
}
\description{
Export estimates, standard errors, degrees of freedom, confidence intervals (CIs) and p-values for the mean coefficients of a linear mixed model.
}
\details{
This function simply calls \code{\link{confint}} with a specific value for the argument \code{column}.
}
