% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.tables.R
\name{model.tables.Wald_lmm}
\alias{model.tables.Wald_lmm}
\title{Statistical Inference for Wald test}
\usage{
\method{model.tables}{Wald_lmm}(x, effects = "Wald", columns, simplify = TRUE, ...)
}
\arguments{
\item{x}{a \code{lmm} object.}

\item{effects}{[character] should the linear contrasts involved in the Wald test be output (\code{"Wald"}),
or the contrast matrix (\code{"contrast"})?}

\item{columns}{[character vector] Columns to be output.
Can be any of \code{"estimate"}, \code{"se"}, \code{"statistic"}, \code{"df"}, \code{"null"}, \code{"lower"}, \code{"upper"}, \code{"p.value"}.}

\item{simplify}{[logical] with argument \code{effects="Wald"}, omit from the output attributes containing additional information (e.g. approximation error made when adjusting p-values).
with argument \code{effects="contrast"} the output will be converted into a matrix (instead of a list of matrix) whenever possible.}

\item{...}{arguments to be passed to \code{\link{confint.Wald_lmm}}.}
}
\description{
Export estimates, standard errors, degrees of freedom, confidence intervals (CIs) and p-values for the mean coefficients of a linear mixed model.
}
\keyword{methods}
