% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.lmm}
\alias{nobs.lmm}
\title{Number of Observations from a Linear Mixed Model}
\usage{
\method{nobs}{lmm}(object, ...)
}
\arguments{
\item{object}{an lmm object.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A vector with 4 elements: \itemize{
\item \code{obs}: the number of repetitions with full data
\item \code{cluster}: the number of clusters with a least one repetition with full data
\item \code{missing.obs}: the number of repetitions with missing data
\item \code{missing.cluster}: the number of cluster with only missing data
}
}
\description{
Extract the number of observations from a Linear Mixed Model
}
\keyword{methods}
