% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs.mlmm}
\alias{nobs.mlmm}
\title{Number of Observations from Multiple Linear Mixed Models}
\usage{
\method{nobs}{mlmm}(object, ...)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A matrix with as many rows as models and 4 columns: \itemize{
\item \code{obs}: the number of repetitions with full data
\item \code{cluster}: the number of clusters with a least one repetition with full data
\item \code{missing.obs}: the number of repetitions with missing data
\item \code{missing.cluster}: the number of cluster with only missing data
}
}
\description{
Extract the number of observations from multiple linear mixed models.
}
\keyword{methods}
