% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialCor.R
\name{partialCor}
\alias{partialCor}
\title{Partial Correlation}
\usage{
partialCor(
  formula,
  data,
  repetition = NULL,
  structure = NULL,
  by = NULL,
  effects = NULL,
  rhs = NULL,
  method = "none",
  df = NULL,
  transform.rho = NULL
)
}
\arguments{
\item{formula}{a formula with in the left hand side the variables for which the correlation should be computed
and on the right hand side the adjustment set. Can also be a list of formula for outcome-specific adjustment set.}

\item{data}{[data.frame] dataset containing the variables.}

\item{repetition}{[formula] Specify the structure of the data: the time/repetition variable and the grouping variable, e.g. ~ time|id.}

\item{structure}{[character] Specify the residual variance-covariance structure.
Without repetitions, either \code{"UN"} or \code{"CS"}.
With repetitions, one of \code{"UN"}, \code{"HLAG"}, \code{"LAG"}, \code{"CS"}.}

\item{by}{[character] variable used to stratified the correlation on.}

\item{effects}{[character or matrix] type of contrast to be used for comparing the correlation parameters. One of \code{"Dunnett"}, \code{"Tukey"}, \code{"Sequen"}, or a contrast matrix.}

\item{rhs}{[numeric vector] right hand side for the comparison of correlation parameters.}

\item{method}{[character] adjustment for multiple comparisons (e.g. \code{"single-step"}).}

\item{df}{[logical] Should a Student's t-distribution be used to model the distribution of the coefficient. Otherwise a normal distribution is used.}

\item{transform.rho}{[character] scale on which perform statistical inference (e.g. \code{"atanh"})}
}
\value{
A data.frame with the estimate partial correlation (rho), standard error, degree of freedom, confidence interval, and p-value (test of no correlation).
When \code{structure="CS"} is used with repeated measurements, a second correlation coefficient (r) is output where the between subject variance has been removed (similar to Bland et al. 1995).
}
\description{
Estimate the partial correlation between two variables where the adjustment set may differ between variables.
}
\details{
Fit a mixed model to estimate the partial correlation with the following variance-covariance pattern:
\itemize{
\item \bold{no repetition}: unstructure or compound symmetry structure for M observations, M being the number of variables on the left hand side (i.e. outcomes).
\item \bold{repetition}: structure for M*T observations where M being the number of variables (typically 2) and T the number of repetitions. 
Can be \code{"UN"}: unstructured (except the off-diagonal containing the correlation parameter),
or \code{"HLAG"}: toeplitz by block with variable and repetition specific variance,
or \code{"LAG"}: toeplitz by block, i.e. correlation depending on the gap between repetitions and specific to each variable,
or \code{"CS"}: compound symmetry by block, i.e. variable specific correlation constant over repetitions (except for the off-diagonal containing the correlation parameter).
}
}
\examples{
#### bivariate (no repetition) ####
## example from ppcor::pcor 
y.data <- data.frame(
  hl=c(7,15,19,15,21,22,57,15,20,18),
  disp=c(0.000,0.964,0.000,0.000,0.921,0.000,0.000,1.006,0.000,1.011),
  deg=c(9,2,3,4,1,3,1,3,6,1),
  BC=c(1.78e-02,1.05e-06,1.37e-05,7.18e-03,0.00e+00,0.00e+00,0.00e+00
,  4.48e-03,2.10e-06,0.00e+00)
)

## ppcor::pcor(y.data)
## estimate
##              hl       disp        deg        BC
## hl    1.0000000 -0.6720863 -0.6161163 0.1148459
## disp -0.6720863  1.0000000 -0.7215522 0.2855420
## deg  -0.6161163 -0.7215522  1.0000000 0.6940953
## BC    0.1148459  0.2855420  0.6940953 1.0000000

## $p.value
##              hl       disp        deg         BC
## hl   0.00000000 0.06789202 0.10383620 0.78654997
## disp 0.06789202 0.00000000 0.04332869 0.49299871
## deg  0.10383620 0.04332869 0.00000000 0.05615021
## BC   0.78654997 0.49299871 0.05615021 0.00000000

set.seed(10)
y.data$gender <- factor(rbinom(10, size = 1, prob = 0.5), labels = c("F","M"))

partialCor(c(hl,disp)~BC+deg, data = y.data)
partialCor(hl + disp~BC+deg, data = y.data)

partialCor(list(hl~BC+deg, disp~BC), data = y.data)
partialCor(list(hl~BC+deg+gender, disp~1), data = y.data)

#### stratified bivariate (no repetition) ####
## partialCor(list(hl~1, disp~1), data = y.data, by = "gender") ## too small dataset

#### bivariate (with repetition) ####
\dontrun{
data(gastricbypassL, package = "LMMstar")
## mean: variable and timepoint specific mean parameter (8)
## variance: variable and timepoint specific variance parameter (8)
## correlation: correlation parameter specific for each variable and time lag (10)
e.cor <- partialCor(weight+glucagonAUC~time, repetition =~time|id,
                    data = gastricbypassL, structure = "LAG")
e.cor
coef(attr(e.cor,"lmm"), effects = "correlation")
if(require(ggplot2)){
autoplot(e.cor)
}

## same except for the mean structure: variable specific mean parameter (2)
e.cor2 <- partialCor(weight+glucagonAUC~time, repetition =~time|id,
                    data = gastricbypassL, structure = "LAG")

## mean: variable and timepoint specific mean parameter (8)
## variance: variable specific variance parameter (2)
## correlation: correlation parameter specific for each variable and some time lag (4)
e.cor3 <- partialCor(weight+glucagonAUC~time, repetition =~time|id,
                     data = gastricbypassL, structure = "CS")
e.cor3
coef(attr(e.cor3,"lmm"), effects = "correlation")
if(require(ggplot2)){
autoplot(e.cor3)
}

}
}
\references{
Bland J M, Altman D G. Statistics notes: Calculating correlation coefficients with repeated observations: Part 1—correlation within subjects BMJ 1995; 310 :446 doi:10.1136/bmj.310.6977.446
}
