% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lmm}
\alias{predict.lmm}
\title{Predicted Mean Value For Multivariate Gaussian Model}
\usage{
\method{predict}{lmm}(object, newdata, level = 0.95, df = !is.null(object$df), ...)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{newdata}{[data.frame] the covariate values for each cluster.}

\item{level}{[numeric,0-1] the confidence level of the confidence intervals.}

\item{df}{[logical] Should a Student's t-distribution be used to model the distribution of the predicted mean. Otherwise a normal distribution is used.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A data.frame with 5 columns:\itemize{
\item \code{estimate}: predicted mean.
\item \code{se}: uncertainty about the predicted mean.
\item \code{df}: degree of freedom
\item \code{lower}: lower bound of the confidence interval of the predicted mean
\item \code{upper}: upper bound of the confidence interval of the predicted mean
}
}
\description{
Predicted Mean Value For Multivariate Gaussian Model
}
\examples{
## simulate data in the long format
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")

## fit Multivariate Gaussian Model
eUN.lmm <- lmm(Y ~ X1 + X2 + X5, repetition = ~visit|id, structure = "UN", data = dL, df = FALSE)

## prediction
predict(eUN.lmm, newdata = data.frame(X1 = 1, X2 = 2, X5 = 3))

## with Student's t-distribution
\dontrun{
predict(eUN.lmm, newdata = data.frame(X1 = 1, X2 = 2, X5 = 3), df = TRUE)
}
}
