% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{print.anova_lmm}
\alias{print.anova_lmm}
\title{Print Multivariate Wald Tests For Linear Mixed Models}
\usage{
\method{print}{anova_lmm}(x, print.null = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{anova_lmm}, i.e. output of anova function on a \code{lmm} object.}

\item{print.null}{[logical] should the null hypotheses be printed in the console?}
}
\description{
Print Multivariate Wald Tests For Linear Mixed Models
}
