% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\alias{profile.lmm}
\title{Display Contour of the log-Likelihood}
\usage{
\method{profile}{lmm}(
  fitted,
  effects = NULL,
  maxpts = 100,
  conf.level = 0.95,
  trace = FALSE,
  plot = TRUE,
  size = c(3, 2, 1),
  linetype = 2,
  shape = 19,
  transform.sigma = "none",
  transform.k = "none",
  transform.rho = "none",
  transform.names = TRUE,
  ...
)
}
\arguments{
\item{fitted}{a \code{lmm} object.}

\item{effects}{[character vector] name of the parameters for which the (restricted) log-likelihood should be evaluated.
Alternatively can be the type of parameters, e.g. \code{"mean"}, \code{"variance"}, \code{"correlation"}, or \code{"all"}.}

\item{maxpts}{[integer, >0] the number of points use to discretize the likelihood.}

\item{conf.level}{[numeric, 0-1] the confidence level of the confidence intervals used to decide about the range of values for each parameter.}

\item{trace}{[logical] Show the progress of the execution of the function.}

\item{plot}{[logical] Should a graphical representation of the results be provided?}

\item{size}{[numeric vector of length 3] Size of the point for the MLE, the width of the line representing the likelihood, and the width of the corresponding quadratic approximation.}

\item{linetype}{[integer, >0] type of line used to represent the quadratic approximation of the likelihood.}

\item{shape}{[integer, >0] type of point used to represent the MLE.}

\item{transform.sigma}{[character] Transformation used on the variance coefficient for the reference level. One of \code{"none"}, \code{"log"}, \code{"square"}, \code{"logsquare"} - see details.}

\item{transform.k}{[character] Transformation used on the variance coefficients relative to the other levels. One of \code{"none"}, \code{"log"}, \code{"square"}, \code{"logsquare"}, \code{"sd"}, \code{"logsd"}, \code{"var"}, \code{"logvar"} - see details.}

\item{transform.rho}{[character] Transformation used on the correlation coefficients. One of \code{"none"}, \code{"atanh"}, \code{"cov"} - see details.}

\item{transform.names}{[logical] Should the name of the coefficients be updated to reflect the transformation that has been used?}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Display the (restricted) log-likelihood when varying one parameter at a time.
}
\details{
WARNING: this differs from profile likelihood (which contrain one parameter and update the other accordingly).
In particular, it should not be used to obtain confidence intervals.
}
\examples{
data(gastricbypassW)
e.lmm <- lmm(weight2 ~ weight1 + glucagonAUC1, data = gastricbypassW)
df.profile <- profile(e.lmm, effects = "all", maxpts = 10)
df.profile <- profile(e.lmm, effects = "all", maxpts = 10, transform.sigma = "log")
}
