% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Proportion of Significant Findings}
\usage{
proportion(object, n.sample, trace, ...)
}
\arguments{
\item{object}{\code{Wald_lmm} object}

\item{n.sample}{[numeric,>=0] number of bootstrap sample used to assess the uncertainty.
If 0, then only the point estimate is computed.}

\item{trace}{[logical] shoudl the execution of the boostrap be trace.}

\item{...}{additional arguments passed to code{confint.Wald_lmm}}
}
\value{
a data.frame with the estimated proportion (estimate column), standard error and confidence interval (when boostrap is used).
}
\description{
Evaluate the proportion of test above the statistical significance level
}
