% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.lmm}
\alias{ranef.lmm}
\title{Estimate Random Effect From a Linear Mixed Model}
\usage{
\method{ranef}{lmm}(object, p = NULL)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{p}{[numeric vector] value of the model coefficients to be used. Only relevant if differs from the fitted values.}

\item{nestingStructure}{[list] output of the \code{.nestingRanef} function.}
}
\description{
Recover the random effects from the variance-covariance parameter of a linear mixed model.
}
\details{
Consider the following mixed model:
\deqn{Y = X\beta + \epsilon}
where \eqn{\Sigma_{\epsilon}}, the variance of \eqn{\epsilon}, has a (possibly stratified) compound symmetry structure.
Denoting by \eqn{I} the identity matirx, this mean that \eqn{\Sigma_{\epsilon} = \sigma^2 I + Z \Sigma_{\eta} Z^T}
where \eqn{\Sigma_{\eta}} is the covariance relative to the design matrix \eqn{Z} (e.g. same student or school). So implicitely we have:
\deqn{Y = X\beta + Z \eta + \varepsilon}
where \eqn{\varepsilon \sim \mathcal{N}(0, \sigma^2 I)}. So we can estimate the random effets via:
\deqn{E[Y|\eta] = Z \Sigma_{\eta} \Omega^{-1} (Y-X\beta)}
}
\examples{
data(gastricbypassL, package = "LMMstar")

## random intercept
e.RI <- lmm(weight ~ time + (1|id), data = gastricbypassL)
ranef(e.RI)

## nested random effects

## crossed random effects

}
\keyword{methods}
