% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{rbind.anova_lmm}
\alias{rbind.anova_lmm}
\title{Linear Hypothesis Testing Across Linear Mixed Models}
\usage{
\method{rbind}{anova_lmm}(model, linfct, ..., sep = ": ")
}
\arguments{
\item{model}{a \code{anova_lmm} object (output of \code{anova} applied to a \code{lmm} object)}

\item{linfct}{a \code{anova_lmm} object (output of \code{anova} applied to a \code{lmm} object)}

\item{...}{possibly other \code{anova_lmm} objects}

\item{sep}{[character] character used to separate the outcome and the covariate when naming the tests.}
}
\description{
Linear hypothesis testing accross linear mixed model.
}
