% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Inference via Resampling for Linear Mixed Model}
\usage{
resample(
  object,
  type,
  effects,
  n.sample = 1000,
  trace = TRUE,
  seed = NULL,
  cl = NULL
)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{type}{[character] should permutation test (\code{"perm"}) or non-parametric bootstrap (\code{"boot"}) be used?}

\item{effects}{[character vector] the variable(s) to be permuted or the effect(s) to be tested via non-parametric bootstrap.}

\item{n.sample}{[integer] the number of samples used.}

\item{trace}{[logical] should the execution of the resampling be traced?}

\item{seed}{[integer] Random number generator (RNG) state used when starting resampling.}

\item{cl}{a cluster object passed to \code{pbapply::pblapply}. On plateform other than Windows can also be an integer indicating the number of child-processes for parallel evaluation.}
}
\description{
Non-parametric bootstrap or permutation test for Linear Mixed Models.
}
\details{
Both approach are carried at the cluster level: \itemize{
\item Bootstrap: sampling with replacement clusters. If a cluster is picked twice then different cluster id is used for each pick.
\item Permutation: permuting covariate values between clusters. This only lead to the null hypothesis when the covariate values are constant within clusters.
}
}
