% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaureNA.R
\name{restaureNA}
\alias{restaureNA}
\title{Restaure NA}
\usage{
restaureNA(object, index.na, level, cluster)
}
\arguments{
\item{object}{results when NA should be added to match the user input.}

\item{index.na}{[integer vector] index of the missing values.}

\item{level}{[character] Should missing observations, indicated by the argument \code{index.na}, be restaured (\code{"obs"})
or should missing clusters, indicated by the discrepancy between the name of the object and the argument \code{cluster}, be restaured (\code{"cluster"}).}

\item{cluster}{[list] list containing the number of cluster (\code{n}), name of the clusters (\code{levels}), and integer associated with each clusters (\code{index}).}
}
\description{
Restaure NA in the user output that have been excluded when fitting the LMM.
}
\keyword{internal}
