% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formula.R
\name{rhs.vars}
\alias{rhs.vars}
\title{Extract Variables in the Right Hand Side of a Formula}
\usage{
rhs.vars(formula)
}
\arguments{
\item{formula}{a formula}
}
\description{
Extract Variables in the Right Hand Side of a Formula
}
\examples{
rhs.vars(~X)
rhs.vars(Y~X)
rhs.vars(Y~X1+X2)
rhs.vars(Y~1)
rhs.vars(Y+Z~1)
rhs.vars(~1|id)
rhs.vars(Y~X|id)
rhs.vars(Y+Z~X|id)
}
