% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot for Continuous Variables}
\usage{
scatterplot(
  data,
  formula,
  columns,
  format = NULL,
  group = NULL,
  transform = NULL,
  labeller = "label_value",
  alpha.point = 1,
  breaks = "Sturges",
  size.bar = NULL,
  density = FALSE,
  alpha.area = 0.3,
  method.cor = "pearson",
  size.text = 10,
  digits = c(3, 2)
)
}
\arguments{
\item{data}{[data.frame] dataset containing the variables to be displayed.}

\item{formula}{[formula] formula indicating the variables to be used (outcome~time|id). Long format only.}

\item{columns}{[character vector] Columns whose numerical values are to be displayed. Wide format only.}

\item{format}{[character] Is the dataset in the long (\code{"long"}) or wide (\code{"wide"}) format?}

\item{group}{[character] optional group variable used to color the points, stratify the histogram/density and correlation.}

\item{transform}{[character or function] optional transformation to be applied on the outcome.}

\item{labeller}{[character] passed to \code{ggplot2::facet_grid} to modify the strip labels.}

\item{alpha.point}{[numeric] the transparency level used to display the points in the scatterplot.}

\item{breaks}{[character or numeric vector] algorithm or values used to create the histogram cells.}

\item{size.bar}{[numeric,>0] width of the bars of the histogram.}

\item{density}{[logical] should the density be displayed instead of an histogram.}

\item{alpha.area}{[numeric, 0-1] the transparency level used to display the area under the density curve.}

\item{method.cor}{[character] estimator of the correlation. Argument passed to \code{stats::cor}.
When \code{NA}, the correlation is not displayed.}

\item{size.text}{[numeric,>0] size of the font used to display the correlation or information about missing values.}

\item{digits}{[numeric of length 2] number of digits used to display the correlation or round the percentage of missing values.}
}
\value{
a ggplot object
}
\description{
Produce a matrix of plot for continuous variables: scatterplots, histograms, correlation and missing values.
Inspired from the \code{ggpairs} function of the R package GGally.
}
\details{
In the long format, the outcome variable contains the numerical values to be displayed.
The time variable will be used to spit outcome and display each split separately or jointly with one other split.
The identifier links the outcome values across time.
}
\examples{
data(gastricbypassL, package = "LMMstar")
gastricbypassL$group <- as.numeric(gastricbypassL$id) \%\% 3
data(gastricbypassW, package = "LMMstar")

## single group (wide or long format)
scatterplot(gastricbypassL, formula = weight~time|id)
scatterplot(gastricbypassW, columns = paste0("weight",1:4))

## tune histogram
scatterplot(gastricbypassL, formula = weight~time|id, breaks = 15)

## transform outcome
scatterplot(gastricbypassL, formula = weight~time|id, transform = "log")

## handling missing values
scatterplot(gastricbypassL, formula = glucagonAUC~time|id)

## coloring per group
gg <- scatterplot(gastricbypassL, formula = weight~time|id, group = "group")
gg
if(require(ggplot2)){
gg + scale_color_manual(values = c("orange","blue","purple"))
}

## only display percentage of NAs
scatterplot(gastricbypassL, formula = glucagonAUC~time|id, method.cor = NA)
scatterplot(gastricbypassL, formula = glucagonAUC~time|id, method.cor = NA,
group = "group", size.text = 5)

## simple scatterplot
scatterplot(gastricbypassW, columns = c("weight2","glucagonAUC1"))

}
