% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score.mlmm}
\alias{score.mlmm}
\title{Extract The Score From Multiple Linear Mixed Models}
\usage{
\method{score}{mlmm}(
  object,
  effects = "contrast",
  p = NULL,
  newdata = NULL,
  ordering = "parameter",
  ...
)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{effects}{[character] By default will output the estimates relative to the hypotheses being tested (\code{"contrast"}).
But can also output all model coefficients (\code{"all"}),
or only coefficients relative to the mean (\code{"mean"} or \code{"fixed"}),
or only coefficients relative to the variance structure (\code{"variance"}),
or only coefficients relative to the correlation structure (\code{"correlation"}).}

\item{p}{[list of numeric vector] list of model coefficients to be used. Only relevant if differs from the fitted values.}

\item{newdata}{[NULL] Not used. For compatibility with the generic method.}

\item{ordering}{[character] should the output be ordered by type of parameter (\code{parameter}) or by model (\code{by}).}

\item{...}{passed to \code{score.lmm}.}
}
\description{
Extract or compute the first derivative of the log-likelihood of each linear mixed model.
}
