% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mlmm.R
\name{summary.mlmm}
\alias{summary.mlmm}
\title{Summary of Multiple Linear Mixed Models}
\usage{
\method{summary}{mlmm}(
  object,
  method = NULL,
  print = NULL,
  hide.data = FALSE,
  hide.fit = FALSE,
  ...
)
}
\arguments{
\item{object}{an \code{mlmm} object, output of \code{mlmm}.}

\item{method}{[character] type of adjustment for multiple comparisons: one of \code{"none"}, \code{"bonferroni"}, \code{"single-step"}, \code{"single-step2"}.}

\item{print}{[logical] should the output be printed in the console.
Can be a vector of length 2 where the first element refer to the global tests and the second to the individual tests.}

\item{hide.data}{[logical] should information about the dataset not be printed.}

\item{hide.fit}{[logical] should information about the model fit not be printed.}

\item{...}{other arguments are passed to \code{\link{summary.anova_lmm}}.}
}
\description{
Estimates, p-values, and confidence intevals for multiple linear mixed models.
}
