% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.partialCor}
\alias{summary.partialCor}
\title{Summary for partial correlation}
\usage{
\method{summary}{partialCor}(object, digits = 3, detail = TRUE, ...)
}
\arguments{
\item{object}{a \code{partialCor} object, output of \code{partialCor}.}

\item{digits}{[integer,>0] number of digits used to display numeric values.}

\item{detail}{[integer,>0] passed to \code{print.confint_lmm}. If above 0.5 also display when a back-transformation has been used.}

\item{...}{other arguments are passed to \code{print.confint_lmm}.}
}
\description{
Display estimated partial correlation and associated p-values and confidence intevals.
}
\keyword{methods}
