% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.lmm}
\title{Summary Output for a Linear Mixed Model}
\usage{
\method{summary}{lmm}(
  object,
  digit = 3,
  level = 0.95,
  print = TRUE,
  ci = TRUE,
  hide.fit = FALSE,
  hide.cor = FALSE,
  hide.var = TRUE,
  hide.sd = FALSE,
  hide.mean = FALSE,
  ...
)
}
\arguments{
\item{object}{[lmm] output of the \code{lmm} function.}

\item{digit}{[integer,>0] number of digit used to display numeric values.}

\item{level}{[numeric,0-1] confidence level for the confidence intervals.}

\item{print}{[logical] should the output be printed in the console.}

\item{ci}{[logical] should the confidence intervals be displayed.}

\item{hide.fit}{[logical] should information about the model fit not be printed.}

\item{hide.cor}{[logical] should information about the correlation structure not be printed.}

\item{hide.var}{[logical] should information about the variance not be printed.}

\item{hide.sd}{[logical] should information about the standard deviation not be printed.}

\item{hide.mean}{[logical] should information about the mean structure not be printed.}

\item{...}{not used. For compatibility with the generic function.}
}
\description{
Summary output for a linear mixed model fitted with \code{lmm}.
This is a modified version of the \code{nlme::summary.gls} function.
}
