% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{swabsL}
\alias{swabsL}
\title{Data From The SWABS Study (Long Format)}
\usage{
data(swabsL)
}
\description{
Data from the swabs study,
where the pneumococcus was studied in 18 families with different space available for the household.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item crowding Space available in the household.
\item family Family serial number
\item name Type of family member.
\item swabs number of times the swab measurement was positive.
}
}
\examples{
\dontrun{
swabsL <- read.table("inst/dataTXT/swabs.txt", header = TRUE, na.string = ".")
swabsL$family <- as.factor(swabsL$family)
swabsL$crowding <- factor(swabsL$crowding, 
            levels = c("uncrow","crow","overcrow"))
swabsL$name <- factor(swabsL$name,
   levels = c("mother","father", "child1","child2","child3"))
swabsL <- swabsL[order(swabsL$crowding,swabsL$family,swabsL$name),]
save(swabsL, file = "data/swabsL.rda")
str(swabsL)
}
}
\references{
TODO
}
\keyword{data}
