% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.lmm.R
\name{tidy.lmm}
\alias{tidy.lmm}
\title{Extract Model Coefficients With Confidence Intervals}
\usage{
\method{tidy}{lmm}(x, conf.int = TRUE, conf.level = 0.95, effects = c("mean"), ...)
}
\arguments{
\item{x}{lmm object.}

\item{conf.int}{[logical] Should the confidence interval be output}

\item{conf.level}{[numeric 0-1] Confidence level of the confidence intervals.}

\item{effects}{[character vector] Type of coefficient to be output.
Can be coefficients relative to the expectation of the outcome (\code{"mean"})
or to the variance-covariance structure of the residuals (\code{"variance"}).}

\item{...}{ignored. For compatibility with the generic method.}
}
\description{
Extract model coefficients with confidence intervals.
}
\examples{
library(broom)
data(gastricbypassL)
e.lmm <- lmm(weight ~ time, covariance = ~visit|id, data = gastricbypassL)
tidy(e.lmm)
tidy(e.lmm, effects = c("mean","variance"))
}
