% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformSummaryTable.R
\name{transformSummaryTable}
\alias{transformSummaryTable}
\title{Apply Transformation to Summary Table}
\usage{
transformSummaryTable(object, transform = NULL)
}
\arguments{
\item{object}{A data.frame with columns estimate, se, lower, upper.}

\item{transform}{the name of a transformation or a function.}
}
\value{
a data.frame
}
\description{
Update summary table according to a transformation, e.g. log-transformtion.
P-values are left unchanged but estimates, standard errors, and confidence intervals are updated.
}
