% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.names.R
\name{variable.names.mlmm}
\alias{variable.names.mlmm}
\title{Variables Involved in Multiple Linear Mixed Models}
\usage{
\method{variable.names}{mlmm}(object, effects = "all", original = TRUE, simplify = TRUE, ...)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{effects}{[character] Should all variable be output (\code{"all"}),
or only those related to the outcome (\code{"outcome"}), mean (\code{"mean"}), variance (\code{"variance"}),
correlation (\code{"correlation"}), time (\code{"time"}), cluster (\code{"cluster"}), strata (\code{"strata"}),
or variable defining the split of the dataset on which a separate linear mixed model is fit (\code{"by"})?}

\item{original}{[logical] Should only the variables present in the original data be output?
When \code{FALSE}, variables internally created are output instead of the original variable for time, cluster, and strata.}

\item{simplify}{[logical] Should the list be converted into a vector if a single \code{effects} is requested?}

\item{...}{not used. For compatibility with the generic function}
}
\value{
A list of character vectors or a character vector.
}
\description{
Extract the variables used to obtain multiple linear mixed models.
}
\keyword{methods}
