% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.mlmm}
\alias{vcov.mlmm}
\title{Extract The Variance-Covariance Matrix From Multiple Linear Mixed Models}
\usage{
\method{vcov}{mlmm}(
  object,
  effects = "contrast",
  method = "none",
  robust = object$args$robust,
  newdata = NULL,
  p = NULL,
  type.information = NULL,
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{effects}{[character] By default will output the estimates relative to the hypotheses being tested (\code{"contrast"}).
But can also output all model coefficients (\code{"all"}),
or only coefficients relative to the mean (\code{"mean"} or \code{"fixed"}),
or only coefficients relative to the variance structure (\code{"variance"}),
or only coefficients relative to the correlation structure (\code{"correlation"}).}

\item{robust}{[logical] Should robust standard errors (aka sandwich estimator) be output instead of the model-based standard errors.
Not feasible for variance or correlation coefficients estimated by REML.}

\item{newdata}{[NULL] Not used. For compatibility with the generic method.}

\item{p}{[list of numeric vector] list of model coefficients to be used. Only relevant if differs from the fitted values.}

\item{transform.sigma}{[character] Transformation used on the variance coefficient for the reference level. One of \code{"none"}, \code{"log"}, \code{"square"}, \code{"logsquare"} - see details.}

\item{transform.k}{[character] Transformation used on the variance coefficients relative to the other levels. One of \code{"none"}, \code{"log"}, \code{"square"}, \code{"logsquare"}, \code{"sd"}, \code{"logsd"}, \code{"var"}, \code{"logvar"} - see details.}

\item{transform.rho}{[character] Transformation used on the correlation coefficients. One of \code{"none"}, \code{"atanh"}, \code{"cov"} - see details.}

\item{...}{passed to \code{vcov.lmm}.}
}
\description{
Extract the variance-covariance matrix of the model coefficients from multiple linear mixed models.
}
