% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.rbindWald_lmm}
\alias{vcov.rbindWald_lmm}
\title{Extract  the Variance-Covariance From Wald Tests for Linear Mixed Models}
\usage{
\method{vcov}{rbindWald_lmm}(
  object,
  effects = "Wald",
  method = "none",
  df = FALSE,
  ordering = NULL,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{rbindWald_lmm} object.}

\item{effects}{[character] should the variance-covariance matrix involved in the Wald tests be output (\code{"contrast"}),
or the variance-covariance matrix of the linear mixed model parameters (\code{"all"})?}

\item{method}{[character vector] type of adjustment for multiple comparisons across the linear contrasts (one of \code{"none"}, \code{"bonferroni"}, ..., \code{"single-step2"}).
Only relevant when \code{effects = "contrast"}.}

\item{df}{[logical] Should degree of freedom, computed using Satterthwaite approximation, for the model parameters be output.
Also output the first derivative of the variance-covariance matrix whenever the argument is stricly greater than 1.}

\item{ordering}{[character] should the output be ordered by name of the linear contrast (\code{"contrast"}) or by model (\code{"model"}).
Only relevant when \code{effects="contrast"}, \code{effects="all"}, or \code{effects="all.original"}.}

\item{simplify}{[logical] should the output be a vector or a list with one element specific to each possible ordering (i.e. contrast or model).}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A matrix with one column and column per parameter.
An attribute \code{"df"} is added when argument df is set to \code{TRUE}, containing a numeric vector with one element per parameter.
An attribute \code{"dVcov"} is added when argument df is greater than 1, containing a 3 dimensional array with with dimension being the number of parameters.
}
\description{
Extract  the variance-covariance matrix from Wald tests applied to a linear mixed models.
}
