% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights.Wald_lmm}
\alias{weights.Wald_lmm}
\title{Extract Weights Used to Pool Estimates}
\usage{
\method{weights}{Wald_lmm}(object, method, effects = "Wald", transform.names = FALSE, ...)
}
\arguments{
\item{object}{a \code{Wald_lmm} object, output of \code{anova.lmm}, or \code{rbind.lmm}, or \code{mlmm}.}

\item{method}{[character] method for combining the estimates, one of \code{"average"}, \code{"pool.se"}, \code{"pool.gls"}, \code{"pool.gls1"}, \code{"pool.rubin"}.}

\item{effects}{[character] should the weights relative to the Wald test be output (\code{"Wald"}),
or the one relative to the linear mixed model parameters (\code{"all"})?}

\item{transform.names}{[logical] should the name of the coefficients be updated to reflect the transformation that has been used?}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
a numeric vector whose elements sum to 1.
}
\description{
Extract weights used to pool estimates.
}
\examples{

data(gastricbypassL)

#### linear mixed model ####
e.lmm <- lmm(glucagonAUC ~ 0 + weight + visit,
             data = gastricbypassL, repetition = ~visit | id)
Wald_mu <- anova(e.lmm, effects = c("visit4 - visit1 = 0",
                                    "visit2 - visit1 = 0",
                                    "visit3 - visit1 = 0"))
weights(Wald_mu, method = c("average","pool.se"))
weights(Wald_mu, method = c("average","pool.se"), effects = "all")

#### multiple linear mixed models ####
set.seed(10)
dL <- sampleRem(250, n.times = 3, format = "long")

e.mlmm <- mlmm(Y~X1+X2+X6, repetition = ~visit|id, data = dL,
               by = "X4", effects = "X1=0", structure = "CS")
weights(e.mlmm, method = c("average","pool.se","pool.gls"))
}
\keyword{methods}
