% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\alias{anova.lmm}
\alias{confint.anova_lmm}
\alias{print.anova_lmm}
\title{Multivariate Wald Tests For Linear Mixed Model}
\usage{
\method{anova}{lmm}(
  object,
  effects = NULL,
  rhs = NULL,
  df = !is.null(object$df),
  ci = FALSE,
  type.object = "lmm",
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  transform.names = TRUE,
  ...
)

\method{confint}{anova_lmm}(object, parm, level = 0.95, method = "single-step", ...)

\method{print}{anova_lmm}(x, level = 0.95, method = "single-step", print.null = FALSE, ...)
}
\arguments{
\item{object}{a \code{lmm} object. Only relevant for the anova function.}

\item{effects}{[character] Should the Wald test be computed for all variables (\code{"all"}),
or only variables relative to the mean (\code{"mean"} or \code{"fixed"}),
or only variables relative to the variance structure (\code{"variance"}),
or only variables relative to the correlation structure (\code{"correlation"}).
Can also be use to specify linear combinations of coefficients, similarly to the \code{linfct} argument of the \code{multcomp::glht} function.}

\item{rhs}{[numeric vector] the right hand side of the hypothesis. Only used when the argument effects is a matrix.}

\item{df}{[logical] Should a F-distribution be used to model the distribution of the Wald statistic. Otherwise a chi-squared distribution is used.}

\item{ci}{[logical] Should a confidence interval be output for each hypothesis?}

\item{type.object}{[character] Set this argument to \code{"gls"} to obtain the output from the gls object and related methods.}

\item{transform.sigma, transform.k, transform.rho, transform.names}{are passed to the \code{vcov} method. See details section in \code{\link{coef.lmm}}.}

\item{...}{Not used. For compatibility with the generic method.}

\item{parm}{Not used. For compatibility with the generic method.}

\item{level}{[numeric, 0-1] nominal coverage of the confidence intervals.}

\item{method}{[character] type of adjustment for multiple comparisons: one of \code{"none"}, \code{"bonferroni"}, \code{"single-step"}.
Not relevant for the global test (F-test or Chi-square test) - only relevant when testing each hypothesis and adjusting for multiplicity.}

\item{x}{an \code{anova_lmm} object. Only relevant for print and confint functions.}

\item{print.null}{[logical] should the null hypotheses be printed in the console?}
}
\value{
A list of matrices containing the following columns:\itemize{
\item \code{null}: null hypothesis
\item \code{statistic}: value of the test statistic
\item \code{df.num}: degrees of freedom for the numerator (i.e. number of hypotheses)
\item \code{df.denom}: degrees of freedom for the denominator (i.e. Satterthwaite approximation)
\item \code{p.value}: p-value.
}
as well as an attribute contrast containing the contrast matrix encoding the linear combinations of coefficients (in columns) for each hypothesis (in rows).
}
\description{
Perform a Wald test testing simultaneously several null hypotheses corresponding to linear combinations of the model paramaters.
}
\details{
By default confidence intervals and p-values are adjusted based on the distribution of the maximum-statistic.
This is refered to as a single-step Dunnett multiple testing procedures in table II of Dmitrienko et al. (2013) and is performed using the multcomp package with the option \code{test = adjusted("single-step")}.
}
\examples{
## simulate data in the long format
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")

## fit Linear Mixed Model
eUN.lmm <- lmm(Y ~ X1 + X2 + X5, repetition = ~visit|id, structure = "UN", data = dL)

## chi-2 test
anova(eUN.lmm, df = FALSE)

## F-test
anova(eUN.lmm)
anova(eUN.lmm, effects = "all")
anova(eUN.lmm, effects = c("X1=0","X2+X5=10"), ci = TRUE)

}
\references{
Dmitrienko, A. and D'Agostino, R., Sr (2013), Traditional multiplicity adjustment methods in clinical trials. Statist. Med., 32: 5172-5218. https://doi.org/10.1002/sim.5990.
}
