% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/plot.R
\name{autoplot.mlmm}
\alias{autoplot.mlmm}
\alias{plot.mlmm}
\title{Graphical Display For Multiple Linear Mixed Model}
\usage{
\method{autoplot}{mlmm}(object, type = "forest", ...)

\method{plot}{mlmm}(x, facet_nrow = NULL, facet_ncol = NULL, labeller = "label_value", ...)
}
\description{
Graphical Display For Multiple Linear Mixed Model
}
\section{Functions}{
\itemize{
\item \code{plot(mlmm)}: Graphical Display For Multiple Linear Mixed Model

}}
