% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufit.R
\name{autoplot.ufit}
\alias{autoplot.ufit}
\title{Display Unique Fitted Values}
\usage{
\method{autoplot}{ufit}(object, x = NULL, sep.strata = ", ")
}
\arguments{
\item{object}{output of the \code{ufit} function.}

\item{x}{[character vector] time variable.}

\item{sep.strata}{[character] symbol used to separate the strata values in the caption}
}
\description{
Display all the possible fitted values among the possible combinaisons of categorical covariables.
Continuous covariates are set to a single value.
}
