% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mlmm}
\alias{coef.mlmm}
\title{Extract Coefficients From Multiple Linear Mixed Models}
\usage{
\method{coef}{mlmm}(
  object,
  effects = "Wald",
  method = "none",
  ordering = NULL,
  backtransform = NULL,
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  transform.names = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{effects}{[character] By default will output the estimates relative to the hypotheses being tested (\code{"Wald"}).
But can also output all model coefficients (\code{"all"}),}

\item{method}{[character vector] should the estimated value for the linear contrasts be output (one of \code{"none"}, \code{"bonferroni"}, ..., \code{"single-step2"})
and/or pooled linear contrast estimate(s) (\code{"average"}, \code{"pool.se"}, \code{"pool.gls"}, \code{"pool.gls1"}, \code{"pool.rubin"}, \code{"p.rejection"})?
Only relevant when \code{effects = "Wald"}.}

\item{ordering}{[character] should the output be ordered by name of the linear contrast (\code{"contrast"}) or by model (\code{"model"}).}

\item{backtransform}{[logical] should the estimate be back-transformed?}

\item{transform.sigma, transform.k, transform.rho}{[character] for internal use (delta-method via estimate).}

\item{transform.names}{[logical] Should the name of the coefficients be updated to reflect the transformation that has been used?
Ignored when \code{effects="Wald"}.}

\item{simplify}{[logical] omit from the output an attribute containing the parameter type/model or contrast term.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A numeric vector
}
\description{
Combine estimated parameters or linear contrasts applied on parameters from group-specific linear mixed models.
}
\keyword{methods}
