% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mlmm}
\alias{coef.mlmm}
\title{Extract Coefficients From a Linear Mixed Model}
\usage{
\method{coef}{mlmm}(object, effects = "contrast", ordering = "parameter", ...)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{effects}{[character] By default will output the estimate for the hypothesis being tests.
But can also output all model coefficients (\code{"all"}),
or only coefficients relative to the mean (\code{"mean"} or \code{"fixed"}),
or only coefficients relative to the variance structure (\code{"variance"}),
or only coefficients relative to the correlation structure (\code{"correlation"}).}

\item{ordering}{[character] should the output be ordered by type of parameter (\code{parameter}) or by model (\code{by}).}

\item{...}{passed to \code{coef.Wald_lmm}.}
}
\description{
Extract coefficients from a linear mixed model.
}
