% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.Wald_lmm}
\alias{coef.Wald_lmm}
\title{Extract Coefficients From Wald Tests for Linear Mixed Model}
\usage{
\method{coef}{Wald_lmm}(
  object,
  type = "coef",
  method = "none",
  backtransform = object$args$backtransform,
  ...
)
}
\arguments{
\item{object}{a \code{Wald_lmm} object.}

\item{type}{[character] Should the coefficients be extracted (\code{"coef"}) or the contrast matrix (\code{"contrast"} or \code{"ls.contrast"}).
\code{"contrast"} will try to simplify the output into a matrix whereas \code{"ls.contrast"} will keep the original format (list of list of matrix).}

\item{backtransform}{[logical] should the estimate, standard error, and confidence interval be back-transformed?}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Extract coefficients from Wald tests applied to a linear mixed model.
}
