% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.rbindWald_lmm}
\alias{confint.rbindWald_lmm}
\title{Confidence Intervals From Combined Wald Tests}
\usage{
\method{confint}{rbindWald_lmm}(
  object,
  parm,
  level = 0.95,
  df = NULL,
  method = NULL,
  columns = NULL,
  ordering = NULL,
  backtransform = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{rbindWald_lmm} object.}

\item{parm}{Not used. For compatibility with the generic method.}

\item{level}{[numeric, 0-1] nominal coverage of the confidence intervals.}

\item{df}{[logical] Should a Student's t-distribution be used to model the distribution of the Wald statistic. Otherwise a normal distribution is used.}

\item{method}{[character vector] type of adjustment for multiple comparisons across the linear contrasts (one of \code{"none"}, \code{"bonferroni"}, ..., \code{"single-step2"})
and/or pooling methods (\code{"average"}, \code{"pool.se"}, \code{"pool.gls"}, \code{"pool.gls1"}, \code{"pool.rubin"}, \code{"p.rejection"})?
Only relevant when \code{effects = "Wald"}.}

\item{columns}{[character vector] Columns to be output.
Can be any of \code{"estimate"}, \code{"se"}, \code{"statistic"}, \code{"df"}, \code{"null"}, \code{"lower"}, \code{"upper"}, \code{"p.value"}.}

\item{ordering}{[character] should the output be ordered by name of the linear contrast (\code{"contrast"}) or by model (\code{"model"}).}

\item{backtransform}{[logical] should the estimates be back-transformed?}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Compute confidence intervals from combined Wald tests applied to a linear mixed models.
}
\details{
Argument \bold{effects}: when evaluating the proportion of rejected hypotheses (\code{effects="p.rejection"})
a \code{"single-step"} method will be used by default to evaluate the critical quantile.
This can be changed by adding adjustment method, e.g. \code{effects=c("bonferronin","p.rejection"}, in the argument.
}
