% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{CS}
\alias{CS}
\title{Compound Symmetry Structure}
\usage{
CS(formula, var.cluster, var.time, type = NULL, group.type = NULL, add.time)
}
\arguments{
\item{formula}{formula indicating on which variable to stratify the residual variance and correlation (left hand side)
and variables influencing the residual variance and correlation (right hand side).}

\item{var.cluster}{[character] cluster variable.}

\item{var.time}{[character] time variable.}

\item{type}{[character] \itemize{
\item \code{"ho"}, \code{"homo"}, or \code{"homogeneous"}: constant variance and covariate-specific correlation.
Analogous to crossed or nested random effects.
\item \code{"he"}, \code{"hetero"}, or \code{"heterogeneous"}: variance and correlation specific to the level of covariates.
Can be seen as more flexible crossed or nested random effects model.
}}

\item{group.type}{[integer vector] grouping of the regressor for the correlation structure.
A constant value corresponds to nested random effects (default) and a regressor-specific value to crossed random effects}

\item{add.time}{not used.}
}
\value{
An object of class \code{CS} that can be passed to the argument \code{structure} of the \code{lmm} function.
}
\description{
Variance-covariance structure where the variance and correlation of the residuals is constant within covariate levels.
Can be stratified on a categorical variable.
The default has no covariate and therefore the variance and correlation are constant within cluster.
}
\details{
A typical formula would be \code{~1}. It will model: \itemize{
\item \eqn{\sigma^2_{j}=\sigma^2}: a variance constant over repetitions.
\item \eqn{\rho_{j,j'}=\rho}: a correlation constant over repetitions.
}
With 4 repetitions (\eqn{(j,j') \in \{1,2,3,4\}^2} and \eqn{j\neq j'}) this leads to the following residuals variance-covariance matrix:
\deqn{ \begin{bmatrix} \sigma^2 & \rho^2\sigma^2 & \rho^2\sigma^2 & \rho^2\sigma^2 \\ \rho^2\sigma^2 & k_2^2\sigma^2 & \rho^2\sigma^2 & \rho^2\sigma^2 \\ \rho^2\sigma^2 & \rho^2\sigma^2 & k_3^2\sigma^2 & \rho^2\sigma^2 \\ \rho^2\sigma^2 & \rho^2\sigma^2 & \rho^2\sigma^2 & k^2_4\sigma^2 \end{bmatrix}}
It is possibly to stratify the covariance pattern on a categorical variable (e.g. sex) using a formula such as \code{sex~1}. This will lead to:
\deqn{ \begin{bmatrix} \sigma_{\text{female}}^2 & \rho^2\sigma^2 & \rho^2\sigma^2 & \rho^2\sigma^2 \\ \rho^2\sigma^2 & k_{2:\text{female}}^2\sigma_{\text{female}}^2 & \rho^2\sigma^2 & \rho^2\sigma^2 \\ \rho^2\sigma^2 & \rho^2\sigma^2 & k_{3:\text{female}}^2\sigma_{\text{female}}^2 & \rho^2\sigma^2 \\ \rho^2\sigma^2 & \rho^2\sigma^2 & \rho^2\sigma^2 & k^2_{4:\text{female}}\sigma_{\text{female}}^2 \end{bmatrix}
\text{and}
\begin{bmatrix} \sigma_{\text{male}}^2 & 0 & 0 & 0 \\ 0 & k_{2:\text{male}}^2\sigma_{\text{male}}^2 & 0 & 0 \\ 0 & 0 & k_{3:\text{male}}^2\sigma_{\text{male}}^2 & 0 \\ 0 & 0 & 0 & k^2_{4:\text{male}}\sigma_{\text{male}}^2 \end{bmatrix}
}
Instead of stratifying one can also make the variance and correlation specific to arbitrary (categorical) variables, e.g.  \code{~sex} (setting \code{add.time} to \code{FALSE}):
\deqn{ \begin{bmatrix} \sigma^2 & \rho_{\text{female}} \sigma^2 & \rho_{\text{female}} \sigma^2 & \rho_{\text{female}} \sigma^2 \\ \rho_{\text{female}} \sigma^2 & \sigma^2 & \rho_{\text{female}} \sigma^2 & \rho_{\text{female}} \sigma^2 \\ \rho_{\text{female}} \sigma^2 & \rho_{\text{female}} \sigma^2 & \sigma^2 & \rho_{\text{female}} \sigma^2 \\ \rho_{\text{female}} \sigma^2 & \rho_{\text{female}} \sigma^2 & \rho_{\text{female}} \sigma^2 & \sigma^2 \end{bmatrix}
\text{and}
\begin{bmatrix} \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 \\ \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 \\ \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 \\ \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \rho_{\text{male}} \sigma^2 k_{\text{male}}^2 & \sigma^2 k_{\text{male}}^2 \end{bmatrix}
}
}
\examples{
## no covariates
CS(~1, var.cluster = "id", var.time = "time")
CS(sex~1, var.cluster = "id", var.time = "time")

## covariates
CS(~time, var.cluster = "id", var.time = "time")
CS(sex~time, var.cluster = "id", var.time = "time")
CS(list(~time,~1), var.cluster = "id", var.time = "time")
CS(list(sex~time,sex~1), var.cluster = "id", var.time = "time")

}
\keyword{multivariate}
