% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{CS}
\alias{CS}
\title{Compound Symmetry Covariance Matrix}
\usage{
CS(formula)
}
\arguments{
\item{formula}{Time and cluster variables. The left hand side of the formula is ignored.}
}
\description{
Compound symmetry covariance matrix
}
\details{
A typical formula would be \code{~time|id}, indicating a variance constant over time and the same correlation between all pairs of times.
}
