% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{blandAltmanL}
\alias{blandAltmanL}
\title{Data From The Bland Altman Study (Long Format)}
\usage{
data(blandAltmanL)
}
\description{
Data From The Bland Altman Study where two methods to measure the peak expiratory flow rate (PEFR) where compared.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item \code{id}: patient identifier.
\item \code{replicate}: index of the measurement (first or second).
\item \code{method}: device used to make the measurement (Wright peak flow meter or mini Wright peak flow meter).
\item \code{pefr}: measurement (peak expiratory flow rate).
}
}
\references{
Bland & Altman, Statistical methods for assessing agreement between two methods of clinical measurement, Lancet, 1986; i: 307-310.
}
\keyword{datasets}
