% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ckdW}
\alias{ckdW}
\title{CKD wide}
\usage{
data(ckdW)
}
\description{
TODO

\itemize{
\item \code{id}: patient identifier.
\item \code{allocation}:
\item \code{sex}:
\item \code{age}:
\item \code{pwv0}:
\item \code{pwv12}:
\item \code{pwv24}:
\item \code{aix0}:
\item \code{aix12}:
\item \code{aix24}:
\item \code{dropout}:
}
}
\references{
TO ADD
}
\keyword{data}
