% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{potassiumSingleL}
\alias{potassiumSingleL}
\title{Data From The Potassium Intake Study (Long Format)}
\usage{
data(potassiumSingleL)
}
\description{
Data from the potassium intake study,
a randomized placebo-controlled crossover study where the effect of potassium supplement (90 mmol/day) on the renin-angiostensin-aldosteron system (RAAS) was assessed.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item \code{id}: patient identifier.
\item \code{sequence}: treatment group to which the patient has been randomized.
\item \code{period}: time period.
\item \code{treatment}: treatment during the time period.
\item \code{auc}: area under the curve of ?? during the time period.
\item \code{bsauc}: ??
\item \code{aldo}: ??
}
}
\references{
Dreier et al. Effect of increased potassium intake on the reninangiotensinaldosterone system and subcutaneous resistance arteries: a randomized crossover study,
Nephrol Dial Transplant (2020) 110. doi: 10.1093/ndt/gfaa114
}
\keyword{data}
