% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sleepL}
\alias{sleepL}
\title{MREG Data from a sleep deprivation experiment}
\usage{
data(sleepL)
}
\description{
Data from sleep deprivation cross-over study where all subjects have scan at baseline and two follow-up.
They received placebo at one follow-up and carvedilol at the other.
Sequence (Baseline-Placebo-Carvedilol or Baseline-Placebo-Carvedilol) is randomized and the current dataset do not contain the Carvedilol measurement.
At each follow-up they under-go two scans: one awake and one asleep. This was not always possible leading to missing data.
The signal measured relates to spectral characteristics of magnetic resonance encephalography (MREG) and has been modified compared to the real data.

\itemize{
\item \code{id}: patient identifier.
\item \code{deprivation}: is the patient sleep deprived?
\item \code{condition}: treatment received: none at baseline and then placebo or carvedilol.
\item \code{vigilance}: is the patient awake or sleeping?
\item \code{signal.34} and \code{signal.98}: modified MREG signal.
}
}
\keyword{datasets}
