% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df.residual.lmm}
\alias{df.residual.lmm}
\title{Residual Degrees-of-Freedom From a Linear Mixed Model.}
\usage{
\method{df.residual}{lmm}(object, ...)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{...}{Passed to \code{residuals.lmm}.}
}
\value{
A numeric value
}
\description{
Estimate the residual degrees-of-freedom from a linear mixed model.
}
\details{
The residual degrees-of-freedom is estimated using the sum of squared normalized residuals.
}
\keyword{methods}
