% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df.residual.lmm}
\alias{df.residual.lmm}
\title{Residuals Degrees of Freedom}
\usage{
\method{df.residual}{lmm}(object, ...)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{...}{Passed to \code{residuals.lmm}.}
}
\value{
A numeric value
}
\description{
Residuals degrees of freedom. Computed as the sum of squared normalized residuals
}
