% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.coef.R
\name{dummy.coef.lmm}
\alias{dummy.coef.lmm}
\title{Marginal Mean Values For Linear Mixed Model}
\usage{
\method{dummy.coef}{lmm}(object, drop = TRUE, ...)
}
\arguments{
\item{object}{a \code{lmm} object.}

\item{drop}{[logical] should combinations of covariates that do no exist in the original dataset be removed?}

\item{...}{arguments passed to \code{emmeans}.}
}
\value{
A data.frame containing the level for which the means have been computed (if more than one),
the estimated mean (\code{estimate}), standard error (\code{se}), degree of freedom (\code{df}), and 95% confidence interval (\code{lower} and \code{upper}).
}
\description{
Compute the marginal mean (via the emmeans package) for each combination of categorical covariates.
When there is no numeric covariate, this outputs all the mean values fitted by the model.
}
