% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{ID}
\alias{ID}
\title{identity Structure}
\usage{
ID(formula, var.time, ...)
}
\arguments{
\item{formula}{formula indicating the time and cluster variables.}

\item{var.time}{[character] name of the time variable.}

\item{...}{not used.}
}
\description{
Variance-covariance structure where the residuals are independent and identically distribution.
}
\details{
A typical formula would be either \code{~1}.
}
\examples{
ID(~1)
ID(~time)
ID(~time+gender)
ID(~time+gender,var.time="time")
ID(gender~time,var.time="time")

}
