% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{ID}
\alias{ID}
\title{identity Structure}
\usage{
ID(formula, var.cluster, var.time, add.time)
}
\arguments{
\item{formula}{formula indicating on which variable to stratify the residual variance (left hand side).}

\item{var.cluster}{[character] cluster variable.}

\item{var.time}{[character] time variable.}

\item{add.time}{not used.}
}
\value{
An object of class \code{IND} that can be passed to the argument \code{structure} of the \code{lmm} function.
}
\description{
Variance-covariance structure where the residuals are independent and identically distributed.
Can be stratified on a categorical variable.
}
\details{
A typical formula would be \code{~1}.
}
\examples{
ID(NULL, var.cluster = "id", var.time = "time")
ID(~1, var.cluster = "id", var.time = "time")
ID(~gender, var.cluster = "id", var.time = "time")
ID(gender~1, var.cluster = "id", var.time = "time")
}
\keyword{multivariate}
