% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.R
\name{iid.mlmm}
\alias{iid.mlmm}
\title{Extract the Influence Function for Multiple Linear Mixed Models}
\usage{
\method{iid}{mlmm}(x, effects = "Wald", ordering = NULL, transform.names = TRUE, ...)
}
\arguments{
\item{x}{an \code{mlmm} object.}

\item{effects}{[character] should the influence function for the linear contrasts involved in the Wald test be output (\code{"Wald"}),
or for the linear mixed model parameters (\code{"all"})?}

\item{ordering}{[character] should the output be ordered by name of the linear contrast (\code{"contrast"}) or by model (\code{"model"}).}

\item{transform.names}{[logical] should the name of the coefficients be updated to reflect the transformation that has been used?
Only relevant when \code{effects="all"}.}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Extract the influence function of linear contrasts applied to an ML or REML estimator of parameters from group-specific linear mixed models.
}
