% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.R
\name{iid.Wald_lmm}
\alias{iid.Wald_lmm}
\title{Extract the Influence Function from Wald Tests}
\usage{
\method{iid}{Wald_lmm}(x, effects = "Wald", ...)
}
\arguments{
\item{effects}{[character] should the influence function for the linear contrasts involved in the Wald test be output (\code{"Wald"}),
or for the linear mixed model parameters (\code{"all"}),
or a logical value indicating whether the influence function has been stored?}

\item{...}{Not used. For compatibility with the generic method.}

\item{object}{a \code{Wald_lmm} object.}
}
\value{
A matrix with one row per observation and one column per parameter (\code{effects="Wald"} or \code{effects="all"}) or a logical value (\code{effects="test"}).
}
\description{
Extract the influence function of linear mixed model parameters involved in the Wald test.
}
