% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{Linear mixed model}
\usage{
lmm(formula, covariance, data, df = FALSE, ...)
}
\arguments{
\item{formula}{[formula] Specify the model for the mean.
On the left hand side the outcome and on the right hand side the covariates affecting the mean value.
E.g. Y ~ Gender + Gene.}

\item{covariance}{[formula] Specify the model for the covariance.
No left hand side. On the right hand side,
either only the grouping variable (when specifying a compound symmetry structure), e.g. ~1|id,
or the time/repetition variable and the grouping variable, e.g. ~ time|id.}

\item{data}{[data.frame] dataset (in the long format) containing the observations.}

\item{df}{[logical] Should the degree of freedom be computed using a Satterthwaite approximation?}

\item{...}{passed to \code{nlme::gls}.}
}
\description{
Fit a linear mixed model using either a compound symmetry structure or an unstructured covariance matrix.
This is essentially an interface to the \code{nlme::gls} function.
}
\examples{
## simulate data in the wide format
library(lava)
m <- lvm(c(Y1,Y2,Y3,Y4) ~ age + gender)
categorical(m, labels = c("male","female")) <- ~gender
transform(m, id~gender) <- function(x){1:NROW(x)}
distribution(m, ~age) <- gaussian.lvm(mean = 50, sd = 10)

set.seed(10)
dW <- lava::sim(m, 1e2)

## move to the long format
name.varying <- paste0("Y",1:4)
dL <- reshape(dW, direction  = "long",
              idvar = c("id","age","gender"),
              varying = name.varying,
              v.names = "Y",
              timevar = "visit")
rownames(dL) <- NULL
dL$visit <- factor(dL$visit,
                   levels = 1:length(name.varying),
                   labels = name.varying)
head(dL)

## fit mixed model
e0.lmm <- lmm(Y ~ visit + age + gender, covariance = ~1|id, data = dL)
summary(e0.lmm)
nlme:::summary.gls(e0.lmm)

e.lmm <- lmm(Y ~ visit + age + gender, covariance = ~visit|id, data = dL)
cat(attr(e.lmm,"code")) ## code used to fit the model
head(attr(e.lmm,"data")) ## data used to fit the model
summary(e.lmm)
}
