% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{Linear mixed model}
\usage{
lmm(
  formula,
  repetition,
  structure,
  data,
  method.fit = NULL,
  df = NULL,
  type.information = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{formula}{[formula] Specify the model for the mean.
On the left hand side the outcome and on the right hand side the covariates affecting the mean value.
E.g. Y ~ Gender + Gene.}

\item{repetition}{[formula] Specify the model for the covariance.
On the right hand side the time/repetition variable and the grouping variable, e.g. ~ time|id.
On the left hand side, a possible stratification variable, e.g. group ~ time|id. In that case the mean structure should only be stratified on this variable using interactions.}

\item{structure}{[character] type of covariance structure, either \code{"CS"} (compound symmetry) or \code{"UN"} (unstructured).}

\item{data}{[data.frame] dataset (in the long format) containing the observations.}

\item{method.fit}{[character] Should Restricted Maximum Likelihoood (\code{"REML"}) or Maximum Likelihoood (\code{"ML"}) be used to estimate the model parameters?}

\item{df}{[logical] Should the degree of freedom be computed using a Satterthwaite approximation?}

\item{type.information}{[character] Should the expected information be computed  (i.e. minus the expected second derivative) or the observed inforamtion (i.e. minus the second derivative).}

\item{debug}{[interger, >0] Show the progress of the execution of the function.}

\item{...}{passed to \code{nlme::gls}.}
}
\description{
Fit a linear mixed model using either a compound symmetry structure or an unstructured covariance matrix.
This is essentially an interface to the \code{nlme::gls} function.
}
\examples{
## simulate data in the long format
set.seed(10)
dL <- sampleRem(100, n.times = 3, format = "long")

## fit mixed model
eCS.lmm <- lmm(Y ~ X1 + X2 + X5, repetition = ~visit|id, structure = "CS", data = dL)

## output
eCS.lmm
summary(eCS.lmm)
summary(eCS.lmm, ci = TRUE)
}
