% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMstar-package.R
\docType{package}
\name{LMMstar-package}
\alias{LMMstar-package}
\title{LMMstar package: Helper functions for handling repeated measurements in R}
\description{
Companion R package for the course "Statistical analysis
of correlated and repeated measurements for health science
researchers" taught by the section of Biostatistics of the University
of Copenhagen. It provides functions for computing summary statistics
and obtainign graphical displays of longitudinal data, as well as for
statistical modeling and statistical inference using multivariate gaussian model.

Currently only two types of multivariate gaussian models are available: \itemize{
\item one with a compound symmetry structure for the residual variance-covariance matrix. This is equivalent to a random intercept model
\item one with a unstructured structure for the residual variance-covariance matrix
}
In addition, it possible to stratify the residual variance-covariance matrix (and the mean) with respect to a categorical variable.
}
