% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMstar.options.R
\name{LMMstar.options}
\alias{LMMstar.options}
\title{Global options for LMMstar package}
\usage{
LMMstar.options(..., reinitialise = FALSE)
}
\arguments{
\item{...}{options to be selected or updated}

\item{reinitialise}{should all the global parameters be set to their default value}
}
\description{
Update or select global options for the LMMstar package.
}
