% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.tables.R
\name{model.tables.lmm}
\alias{model.tables.lmm}
\title{Statistical Inference for Linear Mixed Model}
\usage{
\method{model.tables}{lmm}(x, columns, ...)
}
\arguments{
\item{x}{a \code{lmm} object.}

\item{columns}{[character vector] Columns to be output.
Can be any of \code{"estimate"}, \code{"se"}, \code{"statistic"}, \code{"df"}, \code{"null"}, \code{"lower"}, \code{"upper"}, \code{"p.value"}, \code{"partial.r"}.}

\item{...}{arguments to be passed to the \code{confint} method. Should not contain the argument \code{column}.}
}
\description{
Export estimates, standard errors, degrees of freedom, confidence intervals (CIs) and p-values for the mean coefficients of a linear mixed model.
}
\details{
This function simply calls \code{\link{confint}} with a specific value for the argument \code{column}.
}
