% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.tables.R
\name{model.tables.mlmm}
\alias{model.tables.mlmm}
\title{Statistical Inference and parametrization of Multiple Linear Mixed Model}
\usage{
\method{model.tables}{mlmm}(
  x,
  parm = NULL,
  level = 0.95,
  method = NULL,
  df = NULL,
  columns = NULL,
  backtransform = NULL,
  ordering = "parameter",
  ...
)
}
\arguments{
\item{x}{a \code{mlmm} object.}

\item{method}{[character] type of adjustment for multiple comparisons, one of \code{"none"}, \code{"bonferroni"}, ..., \code{"fdr"}, \code{"single-step"}, \code{"single-step2"}.
and/or method(s) to pool the estimates, one of \code{"average"}, \code{"pool.se"}, \code{"pool.gls"}, \code{"pool.gls1"}, \code{"pool.rubin"}.
Only relevant when \code{effects = "Wald"}.}

\item{...}{arguments to be passed to \code{\link{confint.lmm}}.}

\item{effects}{[character] Should the CIs/p-values for all coefficients be output (\code{"all"}),
or only for mean coefficients (\code{"mean"} or \code{"fixed"}),
or only for variance coefficients (\code{"variance"}),
or only for correlation coefficients (\code{"correlation"}).
Alternatively can be \code{"param"} to output the name and characteristics of each parameter (type, strata, ...).}

\item{simplify}{[logical] omit from the output the backtransform attribute.
Not relevant when the argument \code{effects="param"},}
}
\value{
A \code{data.frame} object.
}
\description{
Combine estimated parameters with their uncertainty (standard errors, degrees-of-freedom, confidence intervals and p-values) from group-specific linear mixed models
or a table describing each parameter (type, associated sigma or k parameter, ...).
}
\details{
When \code{effects} is not \code{"param"}, this function simply calls \code{\link{confint}} with a specific value for the argument \code{column}.
}
\keyword{methods}
