% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ncgsL}
\alias{ncgsL}
\title{Data From National Cooperative Gallstone Study (Long Format)}
\usage{
data(ncgsL)
}
\description{
Data from the National Cooperative Gallstone Study (NCGS),
a randomized study where the level of serum cholesterol was measured at baseline and after intake of high-dose chenondiol (750mg/day) or placebo.
This dataset is in the long format (i.e. one line per measurement).

\itemize{
\item group Treatment group: highdose or placebo.
\item id Patient identifier
\item visit visit index.
\item cholest cholesterol measurement.
\item time time after the start of the study at which the measurement has been done (in month). Treatment is given at 0+.
}
}
\references{
Grundy SM, Lan SP, Lachin J. The effects of chenodiol on biliary lipids and their association with gallstone dissolution in the National Cooperative Gallstone Study (NCGS). J Clin Invest. 1984 Apr;73(4):1156-66. doi: 10.1172/JCI111301.
}
\keyword{data}
