% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mlmm}
\alias{predict.mlmm}
\title{Predicted Outcome Value by Muliple Linear Mixed Model.}
\usage{
\method{predict}{mlmm}(
  object,
  p = NULL,
  newdata = NULL,
  keep.data = FALSE,
  simplify = TRUE,
  ...
)
}
\description{
Estimate the expected outcome conditional on covariates and possibly on other outcomes based on group-specific linear mixed models.
}
