% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{RE}
\alias{RE}
\title{Random Effect Structure}
\usage{
RE(formula, var.cluster, var.time, ranef = NULL, add.time)
}
\arguments{
\item{formula}{formula indicating on which variable to stratify the residual variance and correlation (left hand side)
and variables influencing the residual variance and correlation (right hand side).##'}

\item{var.cluster}{[character] cluster variable.}

\item{var.time}{[character] time variable.}

\item{ranef}{[list] characteristics of the random effects}

\item{add.time}{not used.}
}
\value{
An object of class \code{CS} that can be passed to the argument \code{structure} of the \code{lmm} function.
}
\description{
Variance-covariance structure parametrized via random effects.
Can be stratified on a categorical variable.
}
\details{
A typical formula would be \code{~1}, indicating a variance constant over time and the same correlation between all pairs of times.
}
\examples{
\dontrun{
data(gastricbypassL, package = "LMMstar")
gastricbypassL$sex <- factor(as.numeric(gastricbypassL$id) \%\% 2,
                             labels = c("female", "male"))
gastricbypassL$baseline <- factor(gastricbypassL$time < 0,
                             labels = c("baseline", "follow-up"))

#### Random intercept ####
eRI <- lmm(glucagonAUC ~ visit + (1|id), data = gastricbypassL)
sigma(eRI)
nlme::ranef(eRI, effects = "variance")
model.tables(eRI, effects = "all")

#### Nested random effects ####
eNRI <- lmm(glucagonAUC ~ visit + (1|id/baseline), data = gastricbypassL)
sigma(eNRI)
## nlme::ranef(eNRI, effects = "variance")
model.tables(eNRI, effects = "all")

#### Cross random effects ####
eCRI <- lmm(glucagonAUC ~ visit + (1|id) + (1|visit), data = gastricbypassL)
summary(eCRI)
## nlme::ranef(eCRI, effects = "variance")
model.tables(eCRI, effects = "all") ## to few observations

data(Penicillin, package = "lme4")
eCRI.2 <- lmm(diameter ~ 1 + (1|plate) + (1|sample), data = Penicillin)
nlme::ranef(eCRI.2, effects = "variance")
model.tables(eCRI.2, effects = "all")

}
}
\keyword{multivariate}
