% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repetition.R
\name{repetition}
\alias{repetition}
\title{Number of Repetitions Within Cluster}
\usage{
repetition(
  formula,
  data,
  format = "factor",
  keep.time = TRUE,
  sep = c(":", ".")
)
}
\arguments{
\item{formula}{[formula] Specify the structure of the data: the time/repetition variable and the grouping variable,
e.g. ~1|id, ~ time|id, or ~time+region|id.}

\item{data}{[data.frame] dataset containing the observations.}

\item{format}{[character] the type of the output: a numeric vector (\code{"numeric"}), a character vector (\code{"character"}), or a factor vector (\code{"factor"}).}

\item{keep.time}{[logical] should the value of the time variable at the repetition be kept in the output (e.g. baseline.1, baseline.2, followUp.1 instead of 1,2,3).
Only relevant when argument \code{formula} contain a time/repetition variable and \code{format="character"} or \code{format="factor"}.}

\item{sep}{[character vector of length 2] character strings used to combine time variables (first element) and the name of the time variable with its values (second element).}
}
\value{
A numeric/character/factor vector, depending on argument \code{format}.
}
\description{
Create a vector containing, for each line of the dataset, the number of occurences of the corresponding cluster up to the current line.
Can stratify the number of occurences on one or several variables.
}
\examples{
data(sleepL, package = "LMMstar")
sleepL$dday <- paste0("d",sleepL$day)
sleepL$rep <- repetition(~1|id, data = sleepL)
sleepL$repDay <- repetition(~dday|id, data = sleepL)
sleepL$repDay.num <- repetition(~dday|id, data = sleepL, format = "numeric")
head(sleepL,15)
}
