% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terms.lmm}
\alias{terms.lmm}
\title{Model Terms For Linear Mixed Models}
\usage{
\method{terms}{lmm}(x, ...)
}
\arguments{
\item{x}{a \code{lmm} object}

\item{...}{not used, for compatibility with the generic method.}
}
\value{
An object of class \code{terms} giving a symbolic representation of the mean structure.
}
\description{
Model terms for linear mixed models. Used by \code{multcomp::glht}.
}
