% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{TOEPLITZ}
\alias{TOEPLITZ}
\title{Toeplitz Structure}
\usage{
TOEPLITZ(formula, var.cluster, var.time, heterogeneous = TRUE, add.time)
}
\arguments{
\item{formula}{formula indicating on which variable to stratify the residual variance and correlation (left hand side)
and variables influencing the residual variance and correlation (right hand side).}

\item{var.cluster}{[character] cluster variable.}

\item{var.time}{[character] time variable.}

\item{heterogeneous}{[logical] when covariates are used for the correlation structure, degree of flexibility within covariate.
1 is unstructured, 0.5 is Toeplitz, and 0 is compound symmetry.}

\item{add.time}{Should the default formula (i.e. when \code{NULL}) contain a time effect.}
}
\value{
An object of class \code{TOEPLITZ} that can be passed to the argument \code{structure} of the \code{lmm} function.
}
\description{
Variance-covariance structure for stationnary processes.
Can be stratified on a categorical variable.
}
\details{
A typical formula would be \code{~1}, indicating a variance constant over time and a correlation specific to each gap time.
Can only be at most one covariate for the correlation structure.
}
\examples{
TOEPLITZ(~1, var.cluster = "id", var.time = "time")
TOEPLITZ(gender~1, var.cluster = "id", var.time = "time")
TOEPLITZ(list(~time,~1), var.cluster = "id", var.time = "time")
TOEPLITZ(list(gender~time,gender~1), var.cluster = "id", var.time = "time")

}
