% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.mlmm}
\alias{vcov.mlmm}
\title{Extract The Variance-Covariance Matrix From Multiple Linear Mixed Models}
\usage{
\method{vcov}{mlmm}(
  object,
  effects = "Wald",
  ordering = NULL,
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  transform.names = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{mlmm} object.}

\item{effects}{[character] should the linear contrasts involved in the Wald test be output (\code{"Wald"}),
or the value of the linear mixed model parameters (\code{"all"})?
Can also contain \code{"gradient"} to also output the gradient of the Variance-Covariance matrix.}

\item{ordering}{[character] should the output be ordered by name of the linear contrast (\code{"contrast"}) or by model (\code{"model"}).}

\item{transform.sigma, transform.k, transform.rho}{[character] for internal use (delta-method via estimate).}

\item{transform.names}{[logical] Should the name of the coefficients be updated to reflect the transformation that has been used?
Only relevant when \code{effects="all"}.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A matrix with one column and column per parameter.  \itemize{
\item \code{effects} includes \code{"gradient"}: with an attribute \code{"gradient"} containing a 3 dimensional array with dimension the number of parameters.
}
}
\description{
Extract the variance-covariance matrix of the model coefficients from multiple linear mixed models.
}
\keyword{methods}
