% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.Wald_lmm}
\alias{vcov.Wald_lmm}
\title{Extract the Variance-Covariance Matrix From Wald Tests}
\usage{
\method{vcov}{Wald_lmm}(
  object,
  effects = "Wald",
  df = FALSE,
  transform.sigma = NULL,
  transform.k = NULL,
  transform.rho = NULL,
  transform.names = TRUE,
  ...
)
}
\arguments{
\item{object}{a \code{Wald_lmm} object.}

\item{effects}{[character vector] should the variance-covariance of the linear contrasts involved in the Wald test be output (\code{"Wald"}),
or of the linear mixed model parameters (\code{"all"})?
Can also contain \code{"gradient"} to also output the gradient of the Variance-Covariance matrix.}

\item{df}{[logical] Should degrees-of-freedom be output?}

\item{transform.sigma, transform.k, transform.rho}{[character] for internal use (delta-method via estimate).}

\item{transform.names}{[logical] Should the name of the coefficients be updated to reflect the transformation that has been used?}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A matrix with one column and column per parameter. \itemize{
\item \code{df=TRUE}: with an attribute \code{"df"} containing a numeric vector with one element per parameter. \cr
\item \code{effects} includes \code{"gradient"}: with an attribute \code{"gradient"} containing a 3 dimensional array with dimension the number of parameters.
}
}
\description{
Extract the variance-covariance matrix of the linear contrasts involved in Wald tests.
}
